% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectre_oop.R
\name{as_red_analysis}
\alias{as_red_analysis}
\alias{as_red_analysis.spectre}
\title{Convert an object to a \code{red_analysis} instance.}
\usage{
as_red_analysis(x, ...)

\method{as_red_analysis}{spectre}(x, kdim = 3, skip_last = TRUE, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to the methods.}

\item{kdim}{number of the smallest eigenvectors included used as components
of the \code{red_analysis} object.}

\item{skip_last}{logical, should the last eigenvector be omitted before
drawing the k trailing eigenvectors?}
}
\description{
Convert an object to an instance of \code{\link{red_analysis}} class.
}
\details{
In case of \code{\link{spectre}} objects, the \code{kdim} trailing eigenvectors,
i.e. eigenvectors with the smallest eigenvalues will be used as components
of the \code{red_analysis} objects. They will be sorted in an ascending order,
i.e. the last eigenvector will be named 'comp_1', the one before last
'comp_2' and so on.
}
