% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_dist}
\alias{calculate_dist}
\alias{calculate_weighted_dist}
\alias{get_kernel_info}
\title{Calculate distance.}
\usage{
calculate_dist(data, method)

calculate_weighted_dist(
  data,
  method = "euclidean",
  weights = 1,
  FUN = function(x, y) x + y,
  ...
)

get_kernel_info()
}
\arguments{
\item{data}{a numeric data frame or matrix (\code{calculate_dist()}) or a list
of such objects (for \code{calculate_weighted_dist()}). Row names are preserved as
observation IDs.}

\item{method}{the name of dissimilarity measure (for \code{calculate_dist()})
or a vector of distance nammes (\code{calculate_weighted_dist()}). See:
\code{\link{get_kernel_info}} for available distances.}

\item{weights}{a numeric vector of weights.}

\item{FUN}{a function used to integrate the weighted distances.}

\item{...}{extra arguments passed to \code{FUN}.}
}
\value{
\code{calculate_dist()}: a matrix with the distance statistics;
\code{get_kernel_info()}: a vector with names of available distance measures.
}
\description{
The \code{calculate_dist()} function calculates distances
between observations (rows) of a data
frame or matrix.
\code{calculate_weighted_dist()} computes a matrix of weighted distances
between observations (rows) for a list of numeric data frames or matrices.
See \code{\link{get_kernel_info}} for a vector of available
distance measures.
}
\details{
\code{calculate_dist()} and \code{calculate_weighted_dist()} provide handy wrappers for
\code{\link[philentropy]{distance}}.
The smc (simple matching coefficient) distance is calculated with the
\code{\link[nomclust]{sm}} function.
}
\references{
Drost H-G. Philentropy: Information Theory and Distance Quantification
with R. J Open Source Softw (2018) 3:765. doi:10.21105/joss.00765

Boriah S, Chandola V, Kumar V. Similarity measures for categorical data:
A comparative evaluation. in Society for Industrial and
Applied Mathematics - 8th SIAM International Conference on Data Mining
2008, Proceedings in Applied Mathematics 130, 243–254.
doi:10.1137/1.9781611972788.22

Sulc Z, Cibulkova J, Rezankova H. nomclust: Hierarchical Cluster Analysis
of Nominal Data. (2021)
Available at: https://cran.r-project.org/package=nomclust
}
