% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{center_data}
\alias{center_data}
\title{Mean or median center a data frame.}
\usage{
center_data(data, type = c("mean", "median"), complete_cases = FALSE)
}
\arguments{
\item{data}{a data frame or a tibble. All variables need to be numeric.}

\item{type}{type of the centering, mean (default) or median.}

\item{complete_cases}{logical, should the observations with the complete
variable record only be included as an output?}
}
\value{
a data frame or a tibble.
}
\description{
Median or mean centering of a data frame or tibble. Preserves
the row names.
}
\details{
a wrapper around \code{\link[base]{scale}}. Mean scaling is equal
to canonical Z-score normalization.
}
