% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_numeric}
\alias{check_numeric}
\title{Check for a numeric data frame or a matrix.}
\usage{
check_numeric(object)
}
\arguments{
\item{object}{an object.}
}
\value{
none. Throws exceptions if the object is not a numeric data frame or
a matrix.
}
\description{
Checks if an object is a numeric data frame or a matrix.
}
