% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_utils.R
\name{check_topo}
\alias{check_topo}
\title{Topology error of self-organizing maps.}
\usage{
check_topo(x)
}
\arguments{
\item{x}{a \code{clust_analysis} object.}
}
\value{
An object of the \code{\link{knb}} class with
\code{\link{summary.knb}} and \code{\link{plot.knb}} methods.
}
\description{
Checks for topology errors for consecutive data points.
}
\details{
The procedure of topology error computation is as follows: for each
observation, two nearest self-organizing map (SOM) nodes are identified.
If such nodes are neighbors in the initial layout of SOM prior to data
fitting, correct topology (coded with 0) is returned and an error otherwise
(coded as 1).
}
