% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{clust_analysis}
\alias{clust_analysis}
\title{Clustering analysis object.}
\usage{
clust_analysis(x)
}
\arguments{
\item{x}{a named list, see Details.}
}
\value{
a \code{clust_analysis} object with the elements listed in Details.
}
\description{
Constructs a \code{clust_analysis} class object given a list with
results of a clustering analysis.
}
\details{
A named list with the following elements is required
as the \code{x} argument:
\itemize{
\item \code{data}: a quosure calling the original data set,
\item \code{dist_mtx}: a numeric matrix with the distances between the observations,
\item \code{dist_method}: name of the distance statistic,
\item \code{clust_obj}: the output object of the clustering analysis,
\item \code{clust_fun}: the name of the clustering function or prediction,
\item \code{clust_assignment}: a data frame with the cluster assignment of the
observations. It has to contain the variables \code{observation} and \code{clust_id},
\item \code{dots}: additional arguments passed to the clustering function.
}

The \code{clust_analysis} object can be created for clustering solutions based on
data frames or matrices or user-provided distance matrices (clustering
functions working with dissimilarity objects of the \code{dist} class).
In the later case an instance of the subclass \code{min_analysis} is returned.
As such, the \code{min_analysis} class inherits most of the methods for
\code{clust_analysis} objects. However, some methods requiring source tabular data
like reduction analysis with the genuine data frame, variable importance
or heterologous cross-distances will not be available.
Semi-supervised clustering and cross-validation will be implemented for
the \code{min_analysis} class in the future.
}
