% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{cluster_cv}
\alias{cluster_cv}
\title{Cross validation results.}
\usage{
cluster_cv(x, ...)
}
\arguments{
\item{x}{a named list with elements listed in Details.}

\item{...}{extra arguments, currently none}
}
\value{
an instance of the \code{cluster_cv} class with the elements specified
in Details.
}
\description{
Creates an object of the \code{cluster_cv} class on a top of a list.
}
\details{
The \code{x} argument has to be a list with the following elements:
\itemize{
\item \code{clust_analysis_object} which stores a \code{clust_analysis} or
\code{combi_analysis} object with the global cluster assignment
\item \code{predictions} with a data frame storing out-of-fold predictions with the
variables \code{observation}, \code{fold_clust} (out-of-fold cluster assignment),
\code{global_clust} (cluster assignment in the global clustering structure),
\code{correct} (a logical indicating the out-of-fold - global assignment
concordance) and \code{fold}(fold ID).
\item \code{fold_stats}, a data frame which stores fold means of
accuracy (\code{corr_rate}), classification error (\code{err_rate}), fraction of
explained clustering variance (\code{frac_var}) and silhouette width (\code{sil_width})
\item \code{summary} storing the global means and BCA 95\% confidence intervals listed
in \code{fold_stats}
}
}
