% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{combi_analysis}
\alias{combi_analysis}
\title{Construct a combi_analysis class object.}
\usage{
combi_analysis(x)
}
\arguments{
\item{x}{a named list, see details.}
}
\description{
Constructs a clust_analysis class object given a list storing
results of reduction analysis or self-organizing map and clustering analysis.
}
\details{
A named list with the following elements is required:
clust_analyses (sores red_analysis or clust_analysis objects),
clust_assignment (a tibble with the cluster assignment of the
observations with the observation and clust_id variables).
}
