% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{components.clust_analysis}
\alias{components.clust_analysis}
\alias{components.combi_analysis}
\title{Dimensionality reduction analysis of the analysis data or distance matrix.}
\usage{
\method{components}{clust_analysis}(
  object,
  kdim = NULL,
  red_fun = c("pca", "mds", "umap"),
  with = c("distance", "data"),
  ...
)

\method{components}{combi_analysis}(
  object,
  kdim = NULL,
  red_fun = c("pca", "mds", "umap"),
  with = c("distance", "data"),
  ...
)
}
\arguments{
\item{object}{an object.}

\item{kdim}{number of dimensions. If NULL, kdim is set to the number of
clusters.}

\item{red_fun}{reduction analysis function: 'pca' (PCA), 'mds' (MDS) or
'umap' (UMAP).}

\item{with}{type of the input data for the reduction analysis:
the clustering data ('data') or the matrix of distances ('distance').}

\item{...}{extra arguments passed to \code{\link{reduce_data}}.}
}
\value{
a `red_analysis` object with the component/score table containing
the cluster assignment information ('clust_id' variable).
}
\description{
Performs principal component analysis (PCA), multi-dimensional
scaling (MDS) or uniform manifold approximation and projection (UMAP) of the
analysis data set used for clustering or distance matrix.
}
\details{
See \code{\link{reduce_data}} for the implementation details.
The distance method, relevant for MDS and UMAP. is taken over from the
clust_object. Hence, some distances may crash the analysis with UMAP, see:
\code{\link[umap]{umap.defaults}} for the compatible distances.
For `combi_analysis` objects, the analysis is done for the global clustering,
i.e. assignment of observations to the clusters and not to the SOM nodes.
}
