% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{components}
\alias{components}
\title{Reduce object data dimensions.}
\usage{
components(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\description{
Performs dimensionality reduction analysis with the given
object's data.
}
