% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_clust_folds}
\alias{create_clust_folds}
\title{Create cross-validation folds for clustering.}
\usage{
create_clust_folds(data, k, seed = 1234)
}
\arguments{
\item{data}{the input data: a data frame, matrix or a list of data frames
or matrices.}

\item{k}{number of cross-validaiton folds.}

\item{seed}{an integer or \code{NULL} which specifies the seed for random
number generator.}
}
\value{
a list with two elements: \code{train} and \code{test} with the data subsets
intended for development of the clustering object and predictions,
respectively.
}
\description{
Creates cross-validation folds in the single-layer (data frame or matrix)
or multi-layer clustering data.
}
\details{
The folds are generated with \code{\link[caret]{createFolds}}.
Intended for internal use.
}
