% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_utils.R
\name{cross2dist}
\alias{cross2dist}
\title{Convert cross-distance objects to a distance matrix.}
\usage{
cross2dist(x, zero_diag = TRUE)
}
\arguments{
\item{x}{a \code{cross_dist} object.}

\item{zero_diag}{logical, should diagonals be filled with zeros?}
}
\value{
a \code{dist} object.
}
\description{
Converts a \code{cross_dist} class object to a matrix of mean
cross-distances between the clusters.
}
