% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{cross_dist}
\alias{cross_dist}
\title{Construct a cross_dist object.}
\usage{
cross_dist(
  x,
  type = c("homologous", "heterologous"),
  method = "euclidean",
  x_levels = NULL,
  y_levels = NULL,
  ...
)
}
\arguments{
\item{x}{a named list of cross-distance matrices.}

\item{type}{type of the cross-distances:
`homologous` for comparison of within the same clustering structure or
`heterologous` for comparison of two clustering structures.}

\item{method}{name of the distance metric.}

\item{x_levels}{order of the clusters of the
first cluster/combi analysis object.}

\item{y_levels}{order of the clusters of the
second cluster/combi analysis object.}

\item{...}{extra arguments, currently none defined.}
}
\value{
an object of class `cross-distance` being a list of
cross-distance matrices.
Information on the comparison type and distance
metric are stored as the `type` and `dist_method` attributes.
Information on the cluster order is stored as the
`x_levels` and `y_levels` attributes.
}
\description{
The `cross_dist` class objects a created on the top of a named list
of matrices of cross-distances between clusters.
The list elements have to be named after the compared clusters following
the 'clust1 vs clust2' scheme.
}
