% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_distance.R
\name{cross_distance}
\alias{cross_distance}
\alias{cross_distance.data.frame}
\alias{cross_distance.clust_analysis}
\alias{cross_distance.combi_analysis}
\title{Compute cross-distances between two data frames}
\usage{
cross_distance(x, ...)

\method{cross_distance}{data.frame}(x, y = NULL, method = "euclidean", ...)

\method{cross_distance}{clust_analysis}(x, y = NULL, method = NULL, ...)

\method{cross_distance}{combi_analysis}(x, y = NULL, method = NULL, ...)
}
\arguments{
\item{x}{a data frame, `clust_analysis` or `combi_analysis` object.}

\item{...}{extra arguments passed to methods.}

\item{y}{an object like `x` or NULL (default).}

\item{method}{distance metric name as specified by
\code{\link{get_kernel_info}}.
For `clust_analysis` or `combi_analysis` instances.
If `method` is set to NULL,
the metric name is extracted from the object (distance between observations
for `clust_analysis` and `combi_analysis`, not between the SOM nodes).}
}
\value{
For data frames: a matrix with pairwise distances,
observations of the `x` data frame
are present in rows, observations of the `y` data frame are presented
in columns.
For `clust_analysis` and `combi_analysis` results:
a list of cross-distance matrices of class \code{\link{cross_dist}}
with defined \code{\link{summary.cross_dist}} and
\code{\link{plot.cross_dist}} methods.
}
\description{
Computes cross-distances between two data frames with the same variable sets.
}
\details{
Distances (for available distances, see: \code{\link{get_kernel_info}}) are
computed in a pair-wise manner employing \code{\link[philentropy]{distance}}.
Preserves row names.
If a single data frame is provided, pairwise observations
between the observations are computed with \code{\link{calculate_dist}}.
If a single `clust_analysis` or `combi_analysis` object is provided,
cross-distances between the clusters within the object are computed.
`cross_distance()` is a S3 generic function.
}
\references{
Drost H-G. Philentropy: Information Theory and Distance Quantification
with R. J Open Source Softw (2018) 3:765. doi:10.21105/joss.00765

Sulc Z, Cibulkova J, Rezankova H. nomclust: Hierarchical Cluster
Analysis of Nominal Data. (2021)
Available at: https://cran.r-project.org/package=nomclust
}
