% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{cv}
\alias{cv}
\title{Cross-validate an object.}
\usage{
cv(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\description{
Performs cross-validation of an object.
}
