% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_utils.R
\name{data_node_distance}
\alias{data_node_distance}
\title{Distance between observations and nodes of a self-organizing map.}
\usage{
data_node_distance(x, kNN = NULL)
}
\arguments{
\item{x}{a \code{clust_analysis} object.}

\item{kNN}{optional, number of the nearest neighbor nodes. If \code{NULL}, a numeric
matrix is returned. If provided, a list of nearest nodes named after
observations is returned.}
}
\value{
a numeric matrix with observations in rows and nodes in columns.
If \code{kNN} is specified a list of nearest neighbor nodes for observations.
}
\description{
Computes distances between observations and nodes of a self-organizing map
(SOM) with distance metrics extracted from a SOM clustering object.
}
