% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{dbscan_cluster}
\alias{dbscan_cluster}
\title{Density clustering with DBSCAN.}
\usage{
dbscan_cluster(
  data,
  distance_method = "euclidean",
  eps,
  minPts = 5,
  seed = 1234,
  ...
)
}
\arguments{
\item{data}{a numeric data frame or matrix or a red_analysis object.}

\item{distance_method}{name of the distance metric, see:
\code{\link{get_kernel_info}}.}

\item{eps}{size (radius) of the epsilon neighborhood.}

\item{minPts}{number of minimum points required in the eps neighborhood for
core points (including the point itself).}

\item{seed}{initial setting of the random number generator.}

\item{...}{extra arguments passed to \code{\link[dbscan]{dbscan}}.}
}
\value{
an object of the class \code{\link{clust_analysis}}.
}
\description{
Performs DBSCAN clustering analysis of a numeric data frame,
matrix or the results of a reduction analysis.
}
\details{
Technically, a wrapper around \code{\link[dbscan]{dbscan}}. If a
red_analysis object is provided as the data argument, the observation
component/score table is subjected to clustering.
If a dissimilarity object of class \code{dist} is provided as \code{data}, the function
returns an instance of subclass of the \code{\link{clust_analysis}} class,
the \code{min_analysis} object. Such objects inherit most of the methods of the
fully fledged clustering analysis object. Yet, methods requiring the source
data frame such as reduction analysis with the genuine data, variable
importance analysis or computation of heterologous cross-distances
are not possible. Semi-supervised clustering and cross-validation are not
implemented at the moment for the
\code{min_analysis} objects.
}
\references{
Hahsler M, Piekenbrock M, Doran D. Dbscan: Fast density-based clustering
with R. J Stat Softw (2019) 91:1–30. doi:10.18637/jss.v091.i01
}
