% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis_oop.R
\name{dist.clust_analysis}
\alias{dist.clust_analysis}
\title{Access the distance of a clust_analysis object.}
\usage{
\method{dist}{clust_analysis}(x, ...)
}
\arguments{
\item{x}{a clust_analysis object.}

\item{...}{extra arguments, currently none.}
}
\value{
a numeric matrix.
}
\description{
Retrieves the matrix with distances between the observations.
}
