% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_analysis_oop.R
\name{dist.combi_analysis}
\alias{dist.combi_analysis}
\title{Access the distances of a combi_analysis object.}
\usage{
\method{dist}{combi_analysis}(x, ...)
}
\arguments{
\item{x}{a combi_analysis object.}

\item{...}{extra arguments, currently none.}
}
\value{
a list of numeric matrices.
}
\description{
Retrieves the matrices with distances between the observations
and SOM nodes.
}
