% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis_oop.R
\name{extract.clust_analysis}
\alias{extract.clust_analysis}
\title{Extract features of a clust_analysis object.}
\usage{
\method{extract}{clust_analysis}(
  x,
  type = c("distance", "assignment", "clust_object", "data", "object"),
  ...
)
}
\arguments{
\item{x}{a clust_analysis object.}

\item{type}{the feature name:
'distance' extracts the matrix with distances between the observations,
'data' the data set used for the clust_object generation,
'assignment' assignment of the observations to the clusters,
'clust_object' or 'object' returns the clustering object.}

\item{...}{extra arguments, currently none.}
}
\value{
the requested feature/property.
}
\description{
A general extractor method for accessing properties and features
of a clust_analysis object.
}
