% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cv_oop.R
\name{extract.cluster_cv}
\alias{extract.cluster_cv}
\alias{summary.cluster_cv}
\title{Cross-validation predictions and result summary.}
\usage{
\method{extract}{cluster_cv}(x, type = c("predictions", "fold_stats"), ...)

\method{summary}{cluster_cv}(object, ...)
}
\arguments{
\item{x}{results of clustering cross-validation, e.g. generated with
\code{\link{cv}} for \code{clust_analysis} or \code{combi_analysis} objects.}

\item{type}{the element to be extracted: 'predictions' or 'fold_stats'.}

\item{...}{extra arguments, currently none.}

\item{object}{results of clustering cross-validation, e.g. generated with
\code{\link{cv}} for \code{clust_analysis} or \code{combi_analysis} objects.}
}
\description{
The functions \code{extract.cluster_cv()} and \code{summary.cluster_cv()} retrieve
out-of-fold predictions, fold-means and global means with 95\% confidence
intervals from results of cross-validation.
}
