% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_analysis_oop.R
\name{extract.combi_analysis}
\alias{extract.combi_analysis}
\title{Extract features of a combi_analysis object.}
\usage{
\method{extract}{combi_analysis}(
  x,
  type = c("distance", "assignment", "clust_object", "data", "object")
)
}
\arguments{
\item{x}{a combi_analysis object.}

\item{type}{the feature name:
'distance' extracts the matrix with distances between the observations,
'data' the data set used for the clust_object generation,
'assignment' assignment of the observations to the clusters,
'clust_object' or 'object' returns the clustering object.}
}
\value{
a list with the requested feature for the observation and node clustering
}
\description{
A general extractor method for accessing properties and features
of a combi_analysis object.
}
