% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extract}
\alias{extract}
\title{Extract object features.}
\usage{
extract(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\description{
Extracts requested features from an object.
}
