% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_analysis_oop.R
\name{extract.red_analysis}
\alias{extract.red_analysis}
\title{Extract features of a red_analysis object.}
\usage{
\method{extract}{red_analysis}(
  x,
  type = c("component_tbl", "scores", "loadings", "data", "sdev", "object"),
  ...
)
}
\arguments{
\item{x}{a red_analysis object, created with \code{\link{reduce_data}}.}

\item{type}{the feature name:
'component_tbl' or 'scores' return the component pr score tables for the
observations, 'loadings' retrieves the table of variable loadings (PCA only),
'data' returns the input data frame, 'sdev' returns standard deviations
associated with the components (PCA only), 'object' returns the PCA or UMAP
object.}

\item{...}{extra arguments, currently none.}
}
\value{
the requested feature/property.
}
\description{
A general extractor method for accessing properties and features
of a red_analysis object.
}
