% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuner_oop.R
\name{extract.tuner}
\alias{extract.tuner}
\title{Extract cluster assignment predictions.}
\usage{
\method{extract}{tuner}(
  x,
  type = c("clust_object", "analysis", "stats", "criteria", "best_tune"),
  ...
)
}
\arguments{
\item{x}{a \code{tuner} object.}

\item{type}{feature to be extracted from the object:
\itemize{
\item \code{clust_object} or \code{analysis} returns the \code{clust_analysis} object generated
with the best combination of the tuning parameters
\item \code{stats} extracts quality statistics for combinations of the tuning
parameters
\item \code{criteria} returns a data frame with criteria of selection of the best
combination of the tuning parameters
\item \code{best_tune} extracts a data frame with the best values of the tuning
parameters
}}

\item{...}{extra arguments, currently none.}
}
\value{
a \code{clust_analysis} object.
}
\description{
The function extracts the \code{clust_analysis} object, quality statistics or
the best combination of the tuning parameters.
}
