% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_functions.R
\name{fa}
\alias{fa}
\title{Factor analysis.}
\usage{
fa(data, kdim = 2, ...)
}
\arguments{
\item{data}{a data frame.}

\item{kdim}{dimension number.}

\item{...}{extra arguments passed to \code{\link[stats]{factanal}}.}
}
\description{
Performs factor analysis of the data. A wrapper
around \code{\link[stats]{factanal}}.
}
