% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{ft_hm_single}
\alias{ft_hm_single}
\alias{ft_hm_multi}
\title{Plot clustering feature heat map.}
\usage{
ft_hm_single(x_object, y_object, line_color = NA, discrete_fill = FALSE)

ft_hm_multi(x_object, line_color = NA, discrete_fill = FALSE)
}
\arguments{
\item{x_object}{a \code{clust_analysis} or \code{combi_analysis} object, specifies
clustering of the observations.}

\item{y_object}{a \code{clust_analysis} or \code{combi_analysis} object, specifies
clustering of the features, an optional parameter. Ignored in case of
multi-layer SOM.}

\item{line_color}{color of the line around heat map tiles.}

\item{discrete_fill}{logical, force a discrete fill scale?}
}
\value{
a \code{ggplot} object (single-layer analysis) or a list of
\code{ggplot} objects (multi-layer case).
}
\description{
Plotting of clustering features as a heat map (single layer analysis)
or a list of heat maps (multi-layer SOM). Intended for internal use.
}
