% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{get_clust_tendency}
\alias{get_clust_tendency}
\title{Check clustering tendency of a data set.}
\usage{
get_clust_tendency(data, n, seed = 1234, ...)
}
\arguments{
\item{data}{a data frame, tibble or a matrix. Numeric variables only.}

\item{n}{the number of points selected from sample space which is also the
number of points selected from the given sample (data).}

\item{seed}{initial setting of the random number generator.}

\item{...}{extra arguments passed to
\code{\link[factoextra]{get_clust_tendency}}.}
}
\value{
The values of the Hopkins statistic, p value and a heat map plot.
}
\description{
Check clustering tendency of a data set as compared with a
random data set using Hopkins statistic.
}
\details{
The p value for the Hopkins statistic is calculated based on the
beta distribution of its values. Technically, the function is an enriched
wrapper around \code{\link[factoextra]{get_clust_tendency}}.
}
