% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_dim}
\alias{get_data_dim}
\title{Get dimensions of a data frame or matrix.}
\usage{
get_data_dim(data)
}
\arguments{
\item{data}{a data frame or matrix.}
}
\value{
a list with the requested statistics.
}
\description{
Gets the number of observations and variables of
the given object.
}
