% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sum_sq}
\alias{get_sum_sq}
\title{Calculate clustering sum of squares.}
\usage{
get_sum_sq(dist_mtx, assignment)
}
\arguments{
\item{dist_mtx}{a numeric matrix with the distances.}

\item{assignment}{a data frame with the variable 'clust_id' specifying the
assignment of the observations to the clusters.}
}
\value{
a list with the values of within-cluster ss for the particular
clusters, total within-cluster ss, total ss, total between-cluster ss as well
as the ratio of between-cluster ss to total ss, interpreted as the fraction
of 'explained' clustering variance.
}
\description{
Calculates total, within cluster and between cluster
sum of squares (ss).
}
\details{
The calculation method is independent of the clustering method.
}
