% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis_oop.R
\name{impact.clust_analysis}
\alias{impact.clust_analysis}
\title{Determine clustering feature importance.}
\usage{
\method{impact}{clust_analysis}(x, seed = 1234, .parallel = FALSE)
}
\arguments{
\item{x}{a clust_analysis object.}

\item{seed}{initial setting of the random number generator.}

\item{.parallel}{logical, should the CV be run in parallel? Experimental.}
}
\value{
a data frame of class 'importance'
}
\description{
Determines importance of specific clustering variables by
comparing the fraction of 'explained' clustering variance of the input
clustering object and the object generated with the variable
re-shuffled randomly.
}
