% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{impact}
\alias{impact}
\alias{impact.clust_analysis}
\alias{impact.combi_analysis}
\title{Permutation importance of clustering features.}
\usage{
impact(x, ...)

\method{impact}{clust_analysis}(x, n_iter = 1, seed = 1234, .parallel = FALSE, ...)

\method{impact}{combi_analysis}(x, n_iter = 1, seed = 1234, .parallel = FALSE, ...)
}
\arguments{
\item{x}{a `clust_analysis` object.}

\item{...}{extra arguments, currently none.}

\item{n_iter}{number of iterations, 1 by default.
If the arguments is larger that 1, the function is run multiple times,
which may help at testing variable importance in a more objective way
for different permutations.}

\item{seed}{initial setting of the random number generator.}

\item{.parallel}{logical, should the CV be run in parallel? Experimental.}
}
\value{
a data frame of class \code{\link{importance}} with the defined
\code{\link{plot.importance}} and \code{\link{summary.importance}} methods.
}
\description{
Determines importance of specific clustering variables by
comparing the fraction of 'explained' clustering variance of the input
clustering object and the object generated with the variable
re-shuffled randomly - so called 'permutation' importance.
}
\details{
The permutation importance algorithm is 'blind' or agnostic to the
clustering procedure.
`impact()` is a S3 generic function.
}
\references{
Breiman L. Random forests. Mach Learn (2001) 45:5–32.
doi:10.1023/A:1010933404324
}
