% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{importance}
\alias{importance}
\title{Construct an importance object.}
\usage{
importance(x)
}
\arguments{
\item{x}{a data frame with the following columns:
\itemize{
\item \code{variable} with the names of variables
\item \code{total_wss} with total within-cluster sum of squares
\item \code{total_ss} with total sum of squares
\item \code{between_ss} with between-cluster sum of squares
\item \code{frac_var} with fraction of explained clustering variance
\item \code{frac_diff} with difference in fraction of explained clustering variance
between the genuine clustering analysis and the clustering analysis done with
the variable of interest modified. This variable stores the actual variable
importance metric.
}}
}
\value{
a tibble of the \code{importance} class with the variables listed
in Details.
}
\description{
Constructs an object of class \code{importance} on the top of
a tibble with the clustering variable importance testing results.
}
