% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{importance_cluster}
\alias{importance_cluster}
\title{Determine clustering feature importance.}
\usage{
importance_cluster(
  data,
  clustering_fun = kcluster,
  seed = 1234,
  .parallel = FALSE,
  ...
)
}
\arguments{
\item{data}{a numeric data frame, matrix or a red_analysis object. If a
red_analysis object is provided as the data argument, the observation
component/score table is subjected to clustering.}

\item{clustering_fun}{clustering function. Should return a
\code{clust_analysis} object.}

\item{seed}{initial setting of the random number generator.}

\item{.parallel}{logical, should the CV be run in parallel?}

\item{...}{extra arguments passed to the clustering_fun.}
}
\value{
a data frame with the values of sum of squares and the clustering
variances.
}
\description{
Determines importance of specific clustering variables by
comparing the fraction of 'explained' clustering variance of the input
clustering object and the object generated with the variable
re-shuffled randomly - i.e. so called 'permutation' importance.
}
\references{
Breiman L. Random forests. Mach Learn (2001) 45:5–32.
doi:10.1023/A:1010933404324
}
