% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis_oop.R
\name{is_clust_analysis}
\alias{is_clust_analysis}
\title{Test for the clust_analysis class.}
\usage{
is_clust_analysis(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests if an object belongs to the clust_analysis class.
}
