% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_analysis_oop.R
\name{is_combi_analysis}
\alias{is_combi_analysis}
\title{Test for the combi_analysis class.}
\usage{
is_combi_analysis(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests if an object belongs to the combi_analysis class.
}
