% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_multi_layer}
\alias{is_multi_layer}
\title{Check for multi-layer analysis.}
\usage{
is_multi_layer(x)
}
\arguments{
\item{x}{an instance of the \code{clust_analysis} or \code{combi_analysis} class.}
}
\value{
a logical value
}
\description{
Checks if the \code{clust_analysis} or \code{combi_analysis} object uses
multi-layer SOM.
}
