% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_analysis_oop.R
\name{is_red_analysis}
\alias{is_red_analysis}
\title{Test for the red_analysis class.}
\usage{
is_red_analysis(x)
}
\arguments{
\item{x}{an object to test.}
}
\value{
a logical value.
}
\description{
Tests whwtrer the object is a red_analysis class.
}
