% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{knb}
\alias{knb}
\title{Create a \code{knb} class object.}
\usage{
knb(x, ...)
}
\arguments{
\item{x}{a data frame as specified in Details.}

\item{...}{extra arguments, currently none.}
}
\value{
an instance of the \code{knb} class as described in Details.
}
\description{
Creates an instance of the \code{knb} class at the top of a data frame.
}
\details{
\code{knb} objects store results of testing neighborhood preservation by
self-organizing maps and clustering analyses and inherit most of their
methods from data frames.
The input data frame has to contain the following columns:
\itemize{
\item \code{observation} with observation identifiers
\item \code{clust_id} with assignment of the observations to the clusters
\item \code{kNN_data} with the number of nearest data point neighbors
\item \code{kNN_cluster} with the number of nearest cluster neighbors
}
}
