% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi_supervised.R
\name{map_som}
\alias{map_som}
\alias{map_supersom}
\title{Predict node assignment for self-organizing map.}
\usage{
map_som(object, newdata = NULL, ...)

map_supersom(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a \code{clust_analysis} or a \code{combi_analysis} object.}

\item{newdata}{a numeric data frame, matrix or a \code{red_analysis} object or a
list containing such objects.
If NULL (default), the bare cluster assignment table is returned.}

\item{...}{extra arguments passed to \code{\link[kohonen]{map.kohonen}}.}
}
\value{
a \code{\link{clust_analysis}} object.
}
\description{
Predictions of assignment of the observations to the self-organizing map
SOM nodes are made with the trained SOM neuronal network. The distances,
weights and SOM architecture are extracted from the provided \code{clust_analysis}
or \code{combi_analysis} object.
}
\details{
Uses \code{\link[kohonen]{map.kohonen}} for mapping of the observations onto
the SOM nodes.
If a \code{red_analysis} object is provided as \code{newdata}, the cluster
assignment is predicted for the component/score table. The \code{newdata} input
has to have the same variables as those used for development of the input
cluster_analysis object.
}
\references{
Kohonen T. Self-Organizing Maps. Berlin, Heidelberg: Springer Berlin
Heidelberg (1995). doi:10.1007/978-3-642-97610-0

Wehrens R, Kruisselbrink J. Flexible self-organizing maps in kohonen 3.0.
J Stat Softw (2018) 87:1–18. doi:10.18637/jss.v087.i07
}
