% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_functions.R
\name{mds}
\alias{mds}
\title{Perform multi-dimensional scaling.}
\usage{
mds(data, distance_method = "euclidean", kdim = 2, ...)
}
\arguments{
\item{data}{a data frame.}

\item{distance_method}{name of the distance metric, see:
\code{\link{get_kernel_info}}}

\item{kdim}{dimension number.}

\item{...}{extra arguments passed to \code{\link[stats]{cmdscale}}.}
}
\description{
Performs multi-dimensional scaling of the data. A wrapper around
\code{\link[stats]{cmdscale}}.
}
