% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{min_max}
\alias{min_max}
\title{Minimum/maximum scaling of a data frame.}
\usage{
min_max(data, complete_cases = FALSE)
}
\arguments{
\item{data}{a data frame or a tibble. All variables need to be numeric.}

\item{complete_cases}{logical, should the observations with the complete
variable record only be included as an output?}
}
\value{
a data frame or a tibble.
}
\description{
Min/max scale of a data frame or tibble. Preserves the
row names.
}
