% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis_oop.R
\name{model.frame.clust_analysis}
\alias{model.frame.clust_analysis}
\title{Access the combi_analysis data frames.}
\usage{
\method{model.frame}{clust_analysis}(formula, ...)
}
\arguments{
\item{formula}{a combi_analysis object.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame.
}
\description{
Retrieves the data frame used for clustering analysis.
}
