% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_analysis_oop.R
\name{model.frame.combi_analysis}
\alias{model.frame.combi_analysis}
\title{Access the clust_analysis data frame.}
\usage{
\method{model.frame}{combi_analysis}(formula, ...)
}
\arguments{
\item{formula}{a combi_analysis object.}

\item{...}{extra arguments, currently none.}
}
\value{
a list of data frames.
}
\description{
Retrieves the data frame used for clustering analysis.
}
