% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_analysis_oop.R
\name{model.frame.red_analysis}
\alias{model.frame.red_analysis}
\title{Model frame/data of a red_analysis object.}
\usage{
\method{model.frame}{red_analysis}(formula, ...)
}
\arguments{
\item{formula}{a red_analysis object.}

\item{...}{extra arguments, currently none.}
}
\description{
Extracts the data frame used for construction of a red_analysis
object.
}
