% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis_oop.R
\name{ngroups.clust_analysis}
\alias{ngroups.clust_analysis}
\title{Numbers of observations assigned to the clusters.}
\usage{
\method{ngroups}{clust_analysis}(x, ...)
}
\arguments{
\item{x}{a clust_analysis object.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with the numbers of observations in the clusters
}
\description{
Retrieves the numbers of observations assigned to the clusters.
}
