% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_analysis_oop.R
\name{ngroups.combi_analysis}
\alias{ngroups.combi_analysis}
\title{Numbers of observations assigned to the clusters.}
\usage{
\method{ngroups}{combi_analysis}(x, ...)
}
\arguments{
\item{x}{a combi_analysis object.}

\item{...}{extra arguments, currently none.}
}
\value{
a list of data frames with the numbers of observations in
the clusters
}
\description{
Retrieves the numbers of observations assigned to the clusters.
}
