% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ngroups}
\alias{ngroups}
\title{Get group assignment numbers.}
\usage{
ngroups(x, ...)
}
\arguments{
\item{...}{extra arguments passed to methods.}

\item{object}{an object.}
}
\description{
Returns the numbers of observations assigned to analysis groups
or clusters.
}
