% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_analysis_oop.R
\name{nobs.combi_analysis}
\alias{nobs.combi_analysis}
\title{Number of observations of a combi_analysis object.}
\usage{
\method{nobs}{combi_analysis}(object, ...)
}
\arguments{
\item{object}{a combi_analysis object.}

\item{...}{extra arguments, currently none.}
}
\value{
a list with the numbers of observations and clustering features.
}
\description{
Retrieves the number of observations and SOM nodes used to generate
a combi_analysis object.
}
