% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nobs.default}
\alias{nobs.default}
\title{Get observation number.}
\usage{
\method{nobs}{default}(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{extra arguments passed to \code{\link[stats]{nobs}}.}
}
\description{
Returns the number of complete observations, a wrapper around
\code{\link[stats]{nobs}}.
}
