% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nobs}
\alias{nobs}
\title{Get observation number.}
\usage{
nobs(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{extra arguments passed to methods.}
}
\description{
Returns the number of complete observations.
}
