% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_analysis_oop.R
\name{nobs.red_analysis}
\alias{nobs.red_analysis}
\title{Get the number of observations for a red_analysis object.}
\usage{
\method{nobs}{red_analysis}(object, ...)
}
\arguments{
\item{object}{a red_analysis object.}

\item{...}{extra arguments, currently none.}
}
\description{
Returns the number of complete observations used for
construction of a red_analysis object.
}
