% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectre_oop.R
\name{nobs.spectre}
\alias{nobs.spectre}
\title{Numbers of observations used for spectral decomposition.}
\usage{
\method{nobs}{spectre}(object, ...)
}
\arguments{
\item{object}{an object of class \code{spectre}.}

\item{...}{extra arguments, currently none.}
}
\value{
a numeric value.
}
\description{
Number of observations used for spectral decomposition.
}
