% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_utils.R
\name{node_neighbors}
\alias{node_neighbors}
\title{Find neighbor nodes in the initial layout of a self-organizing map.}
\usage{
node_neighbors(x)
}
\arguments{
\item{x}{a \code{clust_analysis} object.}
}
\value{
a list named after node identifiers with vectors of identifiers of
the nearest nodes.
}
\description{
Identifies nearest nodes in the initial layout of a self-organizing map,
i.e. prior to fitting to the data points.
}
\details{
The grid architecture information is extracted from the \code{kohonen} object.
The function returns NULL and a warning is applied to a non-SOM clustering
analysis object.
}
