% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.R
\name{np}
\alias{np}
\alias{np.clust_analysis}
\alias{np.combi_analysis}
\alias{np.red_analysis}
\title{Neighborhood preservation.}
\usage{
np(x, ...)

\method{np}{clust_analysis}(x, kNN_data = 5, kNN_cluster = NULL, ...)

\method{np}{combi_analysis}(x, kNN_data = 5, kNN_cluster = NULL, type = c("data", "node", "final"), ...)

\method{np}{red_analysis}(x, kNN_data = 5, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}

\item{kNN_data}{number of k-nearest neighbors for data points.}

\item{kNN_cluster}{number of k-nearest neighbors for clusters. If \code{NULL},
the values will be determined automatically. In this case \code{kNN_cluster = 1}
for non-SOM cluster analyses and \code{kNN_cluster = kNN_data} for SOM analyses.}

\item{type}{type of data used for calculation of the neighborhood
preservation. For \code{type = 'data'}, the comparison of neighborhoods is done
between the data points and their SOM node assignment. For \code{type = 'node'},
neighborhoods of the nodes and final clusters are compared.
For \code{type = 'final'}, the analysis is done at the top level, i.e.
neighborhoods of the data points and the final cluster assignment
are evaluated.}
}
\value{
An object of the \code{\link{knb}} class with
\code{\link{summary.knb}} and
\code{\link{plot.knb}} methods.
}
\description{
Checks the fraction of the nearest neighbors of data points, that are
located in the same cluster or in the node of a self-organizing map (SOM).
}
\details{
The function computes fractions of nearest neighbors of each data point that
are preserved after dimensionality reduction or clustering analysis.
In case of reduction analyses, the neighborhoods in the genuine data set and
the data set following dimensionality reduction ('layout') are compared.
In case of simple clustering analyses, the function calculates fractions of
the k-nearest neighbors located in k-nearest clusters of the given data
point. By default, the number of nearest clusters is set to one
(\code{kNN_cluster}), which means that \code{np()} simply checks which fraction of the
nearest data point neighbors is played in the same cluster. In such case,
the neighborhood preservation fraction averaged for the entire clustering
object and particular clusters gives a measure of cluster separation, similar
to \code{\link[cluster]{silhouette}}.
}
\references{
Venna J, Kaski S. Neighborhood preservation in nonlinear projection methods:
An experimental study. Lect Notes Comput Sci (including Subser Lect Notes
Artif Intell Lect Notes Bioinformatics) (2001) 2130:485–491.
doi:10.1007/3-540-44668-0_68#'

Breard GT. Evaluating Self-Organizing Map Quality Measures as Convergence
Criteria Criteria. Open Access Master’s Theses. Paper 1033.
Available at: https://digitalcommons.uri.edu/theses/1033
}
