% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_functions.R
\name{pca}
\alias{pca}
\title{Principal component analysis.}
\usage{
pca(data, kdim = 2, ...)
}
\arguments{
\item{data}{a data frame.}

\item{kdim}{dimension number.}

\item{...}{extra arguments passed to \code{\link[pcaPP]{PCAproj}}.}
}
\description{
Performs principal component analysis of the data. A wrapper
around \code{\link[pcaPP]{PCAproj}}.
}
