% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permute_clust_data}
\alias{permute_clust_data}
\title{Permute clustering variables.}
\usage{
permute_clust_data(data)
}
\arguments{
\item{data}{a data frame, matrix or a list of such objects.}
}
\value{
a list of data frames or data lists.
The first element \code{data} stores the genuine data set.
Each subsequent element is named after the permuted variable.
}
\description{
Creates a series of clustering data sets with consecutive variables
reshuffled by random (permuted). Works for single- and multiple-layer data.
}
