% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_analysis_oop.R
\name{plot.combi_analysis}
\alias{plot.combi_analysis}
\title{Plot selected features of a combi_analysis object.}
\usage{
\method{plot}{combi_analysis}(
  x,
  type = c("diagnostic", "components", "heat_map", "training", "data"),
  cust_theme = ggplot2::theme_classic(),
  jitter_width = 0,
  jitter_height = 0,
  point_alpha = 1,
  ...
)
}
\arguments{
\item{x}{a combi_analysis object.}

\item{type}{the type of plots:
'diagnostic' returns a series of diagnostic plots for the SOM construction
and clustering of the nodes;
'components' plots the results of reduction analysis done with the clustering
data or the distance matrix (see: \code{\link{components.clust_analysis}});
'heat_map' plots the distances between observations and nodes as a heat maps,
'training' plots the mean distance to the SOM winning unit as a function
of the iteration number; 'data' works only if reduction analysis results
were used for clustering and plots the first two comonents/dimensions.}

\item{cust_theme}{a ggplot theme.}

\item{jitter_width}{horizontal jittering of the point in the plots.}

\item{jitter_height}{vertical jittering of the point in the plots.}

\item{point_alpha}{scatter plot's point alpha.}

\item{...}{extra arguments passed to \code{\link{components.clust_analysis}}.}
}
\value{
a ggplot object or a list of ggplot objects, as specified by the
'type' argument
}
\description{
The plotting method for the clust_analysis class. Enables
plotting of the standard diagnostic plots used for the optimal cluster number
determination for the node clustering (dendrogram, WSS- and
silhouette curve), results of the reduction analysis,
heat map of the distances between the observations and SOM nodes as
well as the self-organizing map training process. It is also possible to plot
the first two variables of the clustering data frame, an option which is
attractive, if the clustering of reduction analysis was performed.
}
