% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knb_oop.R
\name{plot.knb}
\alias{plot.knb}
\title{Plot distribution of neighborhood preservation statistic.}
\usage{
\method{plot}{knb}(
  x,
  show_stats = TRUE,
  signif_digits = 2,
  cust_theme = ggplot2::theme_classic(),
  bar_color = "black",
  bar_fill = "steelblue",
  ...
)
}
\arguments{
\item{x}{an object of the \code{\link{knb}} class.}

\item{show_stats}{logical, should the number of observations in the cluster,
percentage of negative silhouette widths and average silhouette statistic
be shown in the plot? Defaults to TRUE.}

\item{signif_digits}{significant digits used for rounding of the statistics
presented in the plot.}

\item{cust_theme}{custom ggplot theme.}

\item{bar_color}{color of the bar line.}

\item{bar_fill}{color of the bars, relevant only for objects without
cluster assignment.}

\item{...}{extra arguments passed to \code{\link[ggplot2]{geom_bar}}.}
}
\value{
a \code{ggplot} class graphic.
}
\description{
Generates a bar plot of neighborhood preservation statistic values for
observations and clusters, similar to a classical silhouette plot (see:
\code{\link{plot.sil_extra}}).
}
