% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectre_oop.R
\name{plot.spectre}
\alias{plot.spectre}
\title{Diagnostic plots for \code{spectre} objects.}
\usage{
\method{plot}{spectre}(
  x,
  type = c("eigenvalues", "eigenvectors"),
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{x}{an object of \code{spectre} class.}

\item{type}{type of the plot as specified in the Details. \code{eigenvalues} by
default.}

\item{cust_theme}{custom \code{ggplot} \code{theme} object.}

\item{...}{extra arguments passed to \code{\link{plot_eigenvalues}} and
\code{\link{plot_eigenvectors}} which control e.g. the number of the trailing
eigenvalues or the eigenvector pair to be plotted.}
}
\value{
a \code{ggplot} graphic object.
}
\description{
Generates scatter plots of eigenvalues and selected eigenvectors.
}
\details{
Two types of plots are generated as specified by the \code{type} argument:
\itemize{
\item \emph{eigenvalues}: the k smallest eigenvalues (by default the trailing 50 ones).
This plot type may be usefull at assessment of the number of eigenvectors
used in other analyses such as clustering.
\item \emph{eigenvectors}: a scatter plot of two selected eigenvectors, be default,
the last and one before last eigenvector.
}
}
