% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuner_oop.R
\name{plot.tuner}
\alias{plot.tuner}
\title{Plot cluster quality statistic values for tuning.}
\usage{
\method{plot}{tuner}(
  x,
  cust_theme = ggplot2::theme_classic(),
  line_alpha = 1,
  point_size = 2,
  ...
)
}
\arguments{
\item{x}{a \code{tuner} object.}

\item{cust_theme}{a custom \code{ggplot} theme.}

\item{line_alpha}{alpha of the lines, applies only to plots of
regularization paths.}

\item{point_size}{size of the data points.}

\item{...}{extra arguments, currently none.}
}
\value{
a list of \code{ggplot} graphic objects.
}
\description{
The function plots quality statistic values (such as explained clustering
variance, silhouette width or mean neighborhood preservation) for
combinations of the tuning parameters.
}
