% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_dendro}
\alias{plot_dendro}
\title{Plot a dendrogram.}
\usage{
plot_dendro(
  clust_str,
  k,
  labels = TRUE,
  cluster_colors = NULL,
  cluster_labels = paste0("Cluster #", 1:k),
  cluster_leg_title = "Cluster",
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  y_lab = NULL,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{clust_str}{an object of the 'hclust' class.}

\item{k}{an integer, the cluster number.}

\item{labels}{logical, should observation labels be presented in the x axis?}

\item{cluster_colors}{colors of the cluster branches, a vector of the length
k + 1. The last color codes for the connecor branches.}

\item{cluster_labels}{cluster names, a text vector of the lenght k.}

\item{cluster_leg_title}{cluster legend title.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{y_lab}{y axis title.}

\item{cust_theme}{custom plot theme, a ggplot2 theme object.}

\item{...}{extra arguments, currently none.}
}
\value{
A ggplot object.
}
\description{
Plots a dendrogram given a clustering object generated by
\code{\link[stats]{hclust}}.
}
\details{
The dendrogram structure is generated with the
\code{\link[stats]{as.dendrogram}} function and graphical layout provided
by \code{\link[dendextend]{color_branches}} and
\code{\link[dendextend]{set}}.
}
