% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralization.R
\name{plot_eigenvalues}
\alias{plot_eigenvalues}
\title{Plot trailing eigenvalues.}
\usage{
plot_eigenvalues(
  spectre_object,
  k = 50,
  plot_title = NULL,
  point_size = 2,
  point_color = "steelblue",
  point_alpha = 0.75,
  cust_theme = ggplot2::theme_classic()
)
}
\arguments{
\item{spectre_object}{an object of \code{spectre} class}

\item{k}{numeric that specifies the number of the least eigenvalues
to be plotted.}

\item{plot_title}{plot title.}

\item{point_size}{size of the data points.}

\item{point_color}{point color.}

\item{point_alpha}{point alpha.}

\item{cust_theme}{custom \code{ggplot} \code{theme} object.}
}
\value{
a \code{ggplot} graphic.
}
\description{
Plots k least eigenvalues.
}
