% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralization.R
\name{plot_eigenvectors}
\alias{plot_eigenvectors}
\title{Scatter plot of selected eigenvectors.}
\usage{
plot_eigenvectors(
  spectre_object,
  x_eigen = NULL,
  y_eigen = NULL,
  plot_title = NULL,
  point_size = 2,
  point_color = "steelblue",
  point_alpha = 0.75,
  point_wjitter = 0,
  point_hjitter = 0,
  cust_theme = ggplot2::theme_classic()
)
}
\arguments{
\item{spectre_object}{an object of \code{spectre} class}

\item{x_eigen}{a numeric. The number of eigenvector to be presented in
the x axis. If \code{NULL}, this is the last eigenvector available in the object.}

\item{y_eigen}{a numeric. The number of eigenvector to be presented in the
y axis. If \code{NULL}, this is the one before last eigenvectro available in
the object.}

\item{plot_title}{plot title.}

\item{point_size}{size of the data points.}

\item{point_color}{point color.}

\item{point_alpha}{point alpha.}

\item{point_wjitter}{jittering width for the data points.}

\item{point_hjitter}{jittering height for the data points.}

\item{cust_theme}{custom \code{ggplot} \code{theme} object.}
}
\value{
a \code{ggplot} graphic object.
}
\description{
Generates a two-dimensional scatter plot for selected eigenvectors.
}
