% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_htk}
\alias{plot_htk}
\title{Generate diagnostic plots for hard threshold KMEANS clustering.}
\usage{
plot_htk(
  x,
  k = NULL,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  cust_theme = ggplot2::theme_classic()
)
}
\arguments{
\item{x}{a \code{clust_analysis} object.}

\item{k}{a numeric vector with the k cluster numbers. If \code{NULL}, it
will be determined automatically.}

\item{plot_title}{title of the plots.}

\item{plot_subtitle}{subtitle of the plots.}

\item{plot_tag}{plot tag.}

\item{cust_theme}{a custom \code{ggplot} theme.}
}
\value{
a list of two elements \code{wss} and \code{silhouette} storing \code{ggplot}
objects with the plots of within-cluster sum of squares and silhouette widths.
}
\description{
Creates the plot of within-cluster sum of squares and mean silhouette width
for the varying cluster numbers.
}
\details{
Corresponds to the output of \code{\link[factoextra]{fviz_nbclust}}.
Intended for internal use.
}
