% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_knn_distance}
\alias{plot_knn_distance}
\title{Plot the mean distance to k-nearest neighbors.}
\usage{
plot_knn_distance(
  diss_obj,
  k,
  eps = NULL,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  cust_theme = ggplot2::theme_classic()
)
}
\arguments{
\item{diss_obj}{a dissimilarity object (e.g. 'dist' class).}

\item{k}{the k number of the nearest neighbors.}

\item{eps}{the distance to be presented in the plot as a horizontal dashed
line. If NULL, the line is hidden.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{cust_theme}{custom plot theme, a ggplot2 theme object.}
}
\value{
A ggplot object.
}
\description{
Plots the sorted (ascending) distances to k-nearest neighbors
(kNN) for each observation in the provided dissimilarity object.
}
\details{
Internally, the mean kNN distances are calculated with the
\code{\link[dbscan]{kNNdist}} function.
}
