% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_point}
\alias{plot_point}
\title{Generate a custom scatter ggplot.}
\usage{
plot_point(
  data,
  x_var,
  y_var,
  fill_var = NULL,
  label_var = NULL,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = x_var,
  y_lab = y_var,
  fill_lab = NULL,
  cust_theme = ggplot2::theme_classic(),
  point_color = "steelblue",
  point_alpha = 1,
  show_segments = FALSE,
  segment_color = "steelblue",
  segment_alpha = 1,
  label_color = point_color,
  txt_color = "black",
  txt_size = 2.5,
  txt_type = c("label", "text"),
  jitter_width = 0,
  jitter_height = 0
)
}
\arguments{
\item{data}{a data frame.}

\item{x_var}{the name of the variable to be presented in the x axis.}

\item{y_var}{the name of the variable to be presented in the y axis.}

\item{fill_var}{optional, the name of the variable coded by the point fill.
If NULL, the point fill is specified by the point_color argument.}

\item{label_var}{optional, the name of the variable to be presented in the
point labels. If NULL, no point labels are displayed.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{x_lab}{x axis title.}

\item{y_lab}{y axis title.}

\item{fill_lab}{fill legend title.}

\item{cust_theme}{custom plot theme, a ggplot2 theme object.}

\item{point_color}{point fill color.}

\item{point_alpha}{point alpha.}

\item{show_segments}{logical, should lines connecting the (0,0) point with
the plot point be displayed?}

\item{segment_color}{color of the connecting lines.}

\item{segment_alpha}{alpha of the connecting lines.}

\item{label_color}{color of the text labels.}

\item{txt_color}{color of the text presented in the labels.}

\item{txt_size}{size of the text presented in the labels.}

\item{txt_type}{type of the displayed text: either as geom_text or geom_label}

\item{jitter_width}{horizontal jittering of the points.}

\item{jitter_height}{vertical jittering of the points.}
}
\value{
a ggplot object.
}
\description{
Generates a simple scatter ggplot.
}
