% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_train_som}
\alias{plot_train_som}
\title{Visualize the SOM training process.}
\usage{
plot_train_som(
  kohonen_object,
  plot_title = NULL,
  plot_subtitle = NULL,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{kohonen_object}{a 'kohonen' class object.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{cust_theme}{custom plot theme, a ggplot2 theme object.}

\item{...}{extra arguments, currently none specified.}
}
\value{
a ggplot object.
}
\description{
Plots the mean distance to the neuron/winning unit as a
function of the iteration number.
}
\references{
Wehrens R, Kruisselbrink J. Flexible self-organizing maps in kohonen 3.0.
J Stat Softw (2018) 87:1–18. doi:10.18637/jss.v087.i07
}
