% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis_oop.R
\name{predict.clust_analysis}
\alias{predict.clust_analysis}
\title{Semi-supervised clustering.}
\usage{
\method{predict}{clust_analysis}(object, newdata = NULL, type = c("class", "propagation"), ...)
}
\arguments{
\item{object}{a clust_analysis object.}

\item{newdata}{a numeric data frame, matrix or a red_analysis object. If NULL
(default), the bare cluster assignment table is returned.}

\item{type}{type of the projection: simple observation matching
('class', default) or kNN label propagation ('propagation').}

\item{...}{extra arguments passed to \code{\link{propagate}}.}
}
\value{
a clust_analysis object.
}
\description{
Projects the cluster assignment onto new data using simple
observation matching or a k-nearest neighbor (kNN) label propagation
algorithm.
}
\details{
For the implementation details, see: \code{\link{propagate}}.
The default distance metric is extracted from the clust_analysis object.
}
