% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_analysis_oop.R
\name{predict.combi_analysis}
\alias{predict.combi_analysis}
\title{Semi-supervised clustering.}
\usage{
\method{predict}{combi_analysis}(object, newdata = NULL, type = c("class", "propagation"), ...)
}
\arguments{
\item{object}{a combi_analysis object.}

\item{newdata}{a numeric data frame, matrix or a red_analysis object. If NULL
(default), the bare cluster assignment table is returned.}

\item{type}{type of the projection: simple observation matching
('class', default) or kNN label propagation ('propagation').}

\item{...}{extra arguments passed to \code{\link{propagate}}.}
}
\value{
a clust_analysis object.
}
\description{
Projects the cluster assignment onto new data using simple
observation matching or a k-nearest neighbor (kNN) label propagation
algorithm.
}
\details{
For the implementation details, see: \code{\link{propagate}}.
The default distance metric is extracted from the combi_analysis object.
The cluster projection is done on the top level, i.e. takes into account the
final assignment of the observations to the clusters and ignoring
the SOM nodes.
}
