% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_analysis_oop.R
\name{predict.red_analysis}
\alias{predict.red_analysis}
\title{Project new data onto a reduction analysis layout.}
\usage{
\method{predict}{red_analysis}(object, newdata, ...)
}
\arguments{
\item{object}{a \code{red_analysis} object, see Details.}

\item{newdata}{a numeric data frame or a numeric matrix with the new
data set.}

\item{...}{extra arguments, currently none.}
}
\value{
an object of the \code{\link{red_analysis}} class.
}
\description{
Predicts reduction analysis scores for a new piece of data with a reduction
algorithm-specific methodology.
}
\details{
Currently implemented only for PCA and UMAP reduction analysis objects.
The method employs internally the function \code{predict()} from the
packages \code{stats} and \code{umap}. For reduction analysis objects
created with other methods, \code{NULL} is returned with a warning.
Of note, specifically for UMAP reduction analysis objects, the method will
work only for few basic distances implemented by the \code{umap} package by
default (Euclidean, Manhattan and cosine).
}
