% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appearance.R
\name{print.clust_analysis}
\alias{print.clust_analysis}
\alias{print.combi_analysis}
\alias{print.cross_dist}
\alias{print.red_analysis}
\alias{print.tuner}
\title{Printing of objects.}
\usage{
\method{print}{clust_analysis}(x, ...)

\method{print}{combi_analysis}(x, ...)

\method{print}{cross_dist}(x, ...)

\method{print}{red_analysis}(x, ...)

\method{print}{tuner}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments, currently none.}
}
\value{
nothing, called for side effects.
}
\description{
Prints a \code{clust_analysis}, \code{combi_analysis}, \code{red_analysis}
or \code{cross_dist} object.
}
