% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_analysis_oop.R
\name{print.combi_analysis}
\alias{print.combi_analysis}
\title{Printing of a clust_analysis object.}
\usage{
\method{print}{combi_analysis}(x, ...)
}
\arguments{
\item{x}{a combi_analysis object.}

\item{...}{extra arguments, currently none.}
}
\description{
Prints a clust_analysis object.
}
