% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_analysis_oop.R
\name{print.red_analysis}
\alias{print.red_analysis}
\title{Print a red_analysis object.}
\usage{
\method{print}{red_analysis}(x, ...)
}
\arguments{
\item{x}{a red_analysis object.}

\item{...}{extra arguments, currently none.}
}
\description{
Prints the red_analysis object.
}
