% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectre_oop.R
\name{print.spectre}
\alias{print.spectre}
\title{Print method for the \code{spectre} class.}
\usage{
\method{print}{spectre}(x, ...)
}
\arguments{
\item{x}{an object of class \code{spectre}.}

\item{...}{extra arguments, currently none.}
}
\value{
none, called for its side effects.
}
\description{
Prints a \code{spectre} object.
}
