% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi_supervised.R
\name{propagate}
\alias{propagate}
\title{Project the cluster assignment with k-NN label propagation.}
\usage{
propagate(
  object,
  newdata = NULL,
  variables = NULL,
  active_variables = FALSE,
  distance_method = NULL,
  kNN = 5,
  simple_vote = TRUE,
  resolve_ties = FALSE,
  kernel_fun = function(x) 1/x,
  detailed = FALSE
)
}
\arguments{
\item{object}{a \code{clust_analysis} or a \code{combi_analysis} object}

\item{newdata}{a numeric data frame, matrix or a \code{red_analysis} object.
If NULL (default), the bare cluster assignment table is returned.}

\item{variables}{an optional vector with names of variables to be used for
the cluster assignment prediction. If \code{NULL} (default), all variables will
be used.}

\item{active_variables}{logical, should the prediction be done with active
variables only? refers only to objects created with hard threshold
regularized algorithms.}

\item{distance_method}{a distance metric, by default it is retrieved from
the input \code{clust_analysis} or \code{combi_analysis} object. For the later, the
distance used for observation clustering is used in the projection.}

\item{kNN}{number of the nearest neighbors.}

\item{simple_vote}{logical, should classical unweighted k-NN classification
be applied? If FALSE, distance-weighted k-NN is used with the provided kernel
function.}

\item{resolve_ties}{logical, should the ties be resolved at random? Applies
only to the simple unweighted voting algorithm.}

\item{kernel_fun}{kernel function transforming the distance into weight.}

\item{detailed}{logical, should a detailed output including the kNN table and
voting scheme be returned. If FALSE, the bare \code{clust_analysis} object with the
predictions is returned.}
}
\value{
a \code{\link{clust_analysis}} object or,
if \code{detailed} = TRUE, a list with the kNN
table and the voting results.
}
\description{
Projects the cluster assignment with a k-nearest neighbor
classifier onto a new data set.
}
\details{
If a \code{red_analysis} object is provided as \code{newdata}, the cluster
assignment is projected onto the component/score table. The \code{newdata} input
has to have the same variables as those used for development of the input
cluster_analysis object. This algorithm is not available for multi-layer
SOM.
}
\references{
Leng M, Wang J, Cheng J, Zhou H, Chen X. Adaptive
semi-supervised clustering algorithm with label propagation.
J Softw Eng (2014) 8:14–22. doi:10.3923/jse.2014.14.22
}
