% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_stats.R
\name{qe}
\alias{qe}
\alias{qe.clust_analysis}
\alias{qe.combi_analysis}
\alias{pbc}
\alias{pbc.clust_analysis}
\alias{pbc.combi_analysis}
\title{Quantization error and population-based convergence.}
\usage{
qe(x, ...)

\method{qe}{clust_analysis}(x, ...)

\method{qe}{combi_analysis}(x, ...)

pbc(x, ...)

\method{pbc}{clust_analysis}(x, ...)

\method{pbc}{combi_analysis}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\value{
\code{qe()} returns a single numeric value. \code{pbc()} returns a list with
two components, \code{variable_stats} with a mean differences, variance ratios,
their lower and upper bounds of 95\% confidence intervals and logical
variables indicating if the convergence was reached.
}
\description{
The functions compute two numeric statistics helpful at assessing convergence
of a self-organizing maps (SOM): quantization error (\code{qe()}) and
population-based convergence (\code{pbc()}).
}
\details{
Quantization error is computed as a sum of distances of data points to their
winning units (node).
The idea behind population-base convergence is to check whether data points
and SOM nodes are drawn from the same population as assessed by comparing
means and variances of the data points and nodes in a variable-wise manner.
The \code{qe()} and \code{pbc()} functions are S3 generics.
}
\references{
\enumerate{
\item Breard GT. Evaluating Self-Organizing Map Quality Measures as Convergence
Criteria Criteria. Open Access Master’s Theses. Paper 1033.
Available at: https://digitalcommons.uri.edu/theses/1033
}

Kohonen T. Self-Organizing Maps. Berlin, Heidelberg: Springer Berlin
Heidelberg (1995). doi:10.1007/978-3-642-97610-0

Hamel L, Ott BH. A Population Based Convergence Criterion for
Self-Organizing Maps. (2012)
}
