% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{red_analysis}
\alias{red_analysis}
\title{Reduction analysis class object.}
\usage{
red_analysis(x)
}
\arguments{
\item{x}{a named list, see Details.}
}
\value{
a \code{red_analysis} object with the elements listed in Details.
}
\description{
Constructs a \code{red_analysis} class object given a list storing
results of a dimensionality reduction analysis.
}
\details{
A named list with the following elements is required as the \code{x} argument:
\itemize{
\item \code{red_obj} with the analysis output,
\item \code{red_fun} name of the reduction function,
\item \code{dist_method} name of the distance metric,
\item \code{component_tbl} a data frame with component/score values for the observations,
\item \code{loadings} a data frame with variable loadings, relevant e.g. for PCA or
factor analysis,
}

\code{data} a quosure calling the original data set.
}
