% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_functions.R
\name{reduce_data}
\alias{reduce_data}
\title{Dimensionality reduction of a data set.}
\usage{
reduce_data(
  data,
  distance_method = "euclidean",
  kdim = 2,
  red_fun = c("pca", "mds", "umap", "fa"),
  ...
)
}
\arguments{
\item{data}{a numeric data frame, a matrix or a distance object
(class \code{dist}).}

\item{distance_method}{name of the distance metric, see:
\code{\link{get_kernel_info}}. Valid only for MDS and UMAP.
For UMAP, the distance is specified by a
\code{\link[umap]{umap.defaults}} object.}

\item{kdim}{dimension number.}

\item{red_fun}{name of the dimensionality reduction function.}

\item{...}{extra arguments passed to \code{\link[pcaPP]{PCAproj}} (PCA),
\code{\link[stats]{cmdscale}} (MDS) and \code{\link[umap]{umap}} (UMAP),
like the \code{\link[umap]{umap.defaults}} object for UMAP,
\code{\link[stats]{factanal}} (FA).}
}
\value{
a \code{\link{red_analysis}} object.
}
\description{
Performs dimensionality reduction of a data frame with principal
component analysis (PCA), multi-dimensional scaling (MDS), Uniform Manifold
Approximation and Projection (UMAP) or factor analysis (FA).
}
\details{
A wrapper around \code{\link[pcaPP]{PCAproj}} (PCA),
\code{\link[stats]{cmdscale}} (MDS), \code{\link[umap]{umap}} (UMAP)
and \code{\link[stats]{factanal}} (FA). Note:
the distances and other UMAP parameters are specified by a
\code{\link[umap]{umap.defaults}} object. Hence, not all distance measures
returned by \code{\link{get_kernel_info}} are available for UMAP computation.
}
\references{
McInnes L, Healy J, Melville J. UMAP: Uniform Manifold Approximation and
Projection for Dimension Reduction. (2018)
Available at: https://arxiv.org/abs/1802.03426v3

Croux C, Filzmoser P, Oliveira MR. Algorithms for Projection-Pursuit robust
principal component analysis. Chemom Intell Lab Syst (2007) 87:218–225.
doi:10.1016/j.chemolab.2007.01.004

BARTLETT MS. THE STATISTICAL CONCEPTION OF MENTAL FACTORS. Br J Psychol
Gen Sect (1937) 28:97–104. doi:10.1111/j.2044-8295.1937.tb00863.x
}
