% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{rename.clust_analysis}
\alias{rename.clust_analysis}
\alias{rename.combi_analysis}
\title{Set cluster names.}
\usage{
\method{rename}{clust_analysis}(.data, nm, ...)

\method{rename}{combi_analysis}(.data, nm, ...)
}
\arguments{
\item{.data}{a \code{clust_analysis} or \code{combi_analysis} object.}

\item{nm}{a named character vector with the new names as elements and old
cluster names as names.}

\item{...}{extra arguments, currently none.}
}
\value{
a \code{\link{clust_analysis}} or \code{\link{combi_analysis}} object.
}
\description{
Sets custom cluster names.
}
\details{
The package's clustering functions of the clustTools package
name clusters with integer numbers by default. This method poses a handy
tool to set custom cluster names with a named character vector.
The cluster order (i.e. vector levels) is defined by the order of
the naming vector's elements.
}
