% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{set_rownames}
\alias{set_rownames}
\title{Set row names.}
\usage{
set_rownames(data, row_names = as.character(1:nrow(data)))
}
\arguments{
\item{data}{a data frame or a tibble.}

\item{row_names}{a character vector of the proper length.}
}
\value{
a data frame.
}
\description{
Sets row names in a data frame.
}
\details{
a tibble is silently converted to a data frame.
}
