% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{sil_extra}
\alias{sil_extra}
\title{Generate a sil_extra object.}
\usage{
sil_extra(x, assignment)
}
\arguments{
\item{x}{an object of the \code{\link[cluster]{silhouette}} class.}

\item{assignment}{an data frame with the `clust_id` and `observation`
columns defining the cluster assignment, e.g. obtained by the
\code{\link{extract}} function applied to a `clust_analysis` or
`combi_analysis` object.}
}
\value{
an object of the `sil_extra` class. Technically, a tibble with
the observation ID, cluster name, neighbor cluster name and silhouette width.
}
\description{
Extends the \code{\link[cluster]{silhouette}} object by
cluster order and names.
}
\details{
The `sil_extra` class has \code{\link{summary.sil_extra}}
and \code{\link{plot.sil_extra}}
methods compatible with the tidyverse environment.
}
