% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{som_cluster}
\alias{som_cluster}
\title{Self-organizing maps.}
\usage{
som_cluster(
  data,
  distance_method = "euclidean",
  xdim = 5,
  ydim = 4,
  topo = c("hexagonal", "rectangular"),
  neighbourhood.fct = c("gaussian", "bubble"),
  toroidal = FALSE,
  seed = 1234,
  ...
)
}
\arguments{
\item{data}{a numeric data frame or matrix or a \code{red_analysis} object or
a list of such objects.}

\item{distance_method}{a character vector with names of the distance metric.
It has to be of the same length as \code{data}, i.e. single name for a single
object and a vector for a list. See: \code{\link{get_kernel_info}} for
available distances.}

\item{xdim}{x dimension of the SOM grid,
see: \code{\link[kohonen]{somgrid}} for details.}

\item{ydim}{y dimension of the SOM grid,
#' see: \code{\link[kohonen]{somgrid}} for details.}

\item{topo}{SOM grid topology, see: \code{\link[kohonen]{somgrid}}
for details. 'hexagonal' for default.}

\item{neighbourhood.fct}{neighborhood function, 'gaussian' for default.}

\item{toroidal}{logical, should toroidal grid be used?}

\item{seed}{initial setting of the random number generator.}

\item{...}{extra arguments passed to \code{\link[kohonen]{som}} or
\code{\link[kohonen]{supersom}} such as the learning rate, iteration
number, algorithm mode or layer weights.}
}
\value{
an object of the class \code{\link{clust_analysis}}. In case of
a multi-layer SOM (\code{data} as a list), the distance matrix corresponds
returned within the \code{clust_analysis} object is the matrix computed by
\code{\link[kohonen]{object.distances}}, which is a weighted distance matrix
over all layers for observations.
}
\description{
Performs self-organizing map (SOM) clustering of a numeric data
frame, matrix or the results of a reduction analysis with the single or
multi-layered Kohonen's on-line or batch algorithm.
}
\details{
Technically, a wrapper around \code{\link[kohonen]{som}} and
\code{\link[kohonen]{supersom}}.
The input can be a single object or a list of numeric data frames,
numeric matrices or \code{red_analysis} objects - combinations are allowed!
The user may specify different distances for each of the elements
of the data list, which allows to handle different types of variables
(e.g. a binary data set handled with Tanimoto distance and a numeric data
set handled with Euclidean distance) and specify weights for each layer
determining how they contribute to the node (cluster) assignment.
If any of the data list elements is a \code{red_analysis} object, its
component/score table will be used for clustering.
Note, in order to make use of the full set of distance measures,
the package 'somKernels' need to be installed and loaded.
}
\references{
Kohonen T. Self-Organizing Maps. Berlin, Heidelberg:
Springer Berlin Heidelberg (1995). doi:10.1007/978-3-642-97610-0

Wehrens R, Kruisselbrink J. Flexible self-organizing maps in kohonen
3.0. J Stat Softw (2018) 87:1–18. doi:10.18637/jss.v087.i07
}
