% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_functions.R
\name{som_reduce}
\alias{som_reduce}
\title{Dimensionality reduction with self-organizing maps.}
\usage{
som_reduce(
  data,
  distance_method = "euclidean",
  xdim = 5,
  ydim = 4,
  topo = c("hexagonal", "rectangular"),
  neighbourhood.fct = c("gaussian", "bubble"),
  toroidal = FALSE,
  seed = 1234,
  ...
)
}
\arguments{
\item{data}{a numeric data frame or matrix.}

\item{distance_method}{name of the distance metric, see:
\code{\link{get_kernel_info}}.}

\item{xdim}{x dimension of the SOM grid,
see: \code{\link[kohonen]{somgrid}} for details.}

\item{ydim}{y dimension of the SOM grid,
#' see: \code{\link[kohonen]{somgrid}} for details.}

\item{topo}{SOM grid topology, see: \code{\link[kohonen]{somgrid}}
for details. 'hexagonal' for default.}

\item{neighbourhood.fct}{neighborhood function, 'gaussian' for default.}

\item{toroidal}{logical, should toroidal grid be used?}

\item{seed}{initial setting of the random number generator.}

\item{...}{extra arguments passed to \code{\link[kohonen]{som}} or
\code{\link[kohonen]{supersom}} such as the learning rate, iteration
number, algorithm mode or layer weights.}
}
\value{
a \code{\link{red_analysis}} object with, among others,
\code{\link{plot.red_analysis}}, \code{\link{summary.red_analysis}} and
\code{\link{predict.red_analysis}} methods.
}
\description{
The \code{som_reduce()} function applies self-organizing maps (SOM) as a classical
dimensionality reduction method, i.e. to reduce the number of variables.
}
\details{
In contrast to \code{\link{som_cluster}}, which reduces the observation
number (observation -> nodes), \code{som_reduce()} diminishes the variable number
by representing associated variables as SOM nodes or 'meta-variables'.
The data frame of such meta-variables (to be strict, codebook vectors of
node positions) is stored as the score/component table of the \code{red_analysis}
object returned by the function.
}
