% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{spectre}
\alias{spectre}
\title{Create a \code{spectre} object.}
\usage{
spectre(x)
}
\arguments{
\item{x}{a list with elements specified in Details.}
}
\value{
an instance of the \code{spectre} class as described in Details.
}
\description{
Creates an instance of \code{spectre} class.
}
\details{
Technically, a \code{spectre} object is a list with the following components:
\itemize{
\item \code{degrees}: a numeric vector with degrees of the graph nodes stored in the
diagonal of the degree matrix
\item \code{eigen_values}: a numeric vector of eigenvalues sorted from the largest to
the smallest one
\item \code{eigen_vectors}: a numeric matrix whose rows represent the observations in
he initial data set and columns representing the eigenvectors.
The eigenvectors are sorted by their einegvalues: the largest come first
}
}
