% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis_oop.R, R/combi_analysis_oop.R
\name{summary.clust_analysis}
\alias{summary.clust_analysis}
\alias{summary.combi_analysis}
\title{Quality control of clustering solutions.}
\usage{
\method{summary}{clust_analysis}(object, ...)

\method{summary}{combi_analysis}(object, ...)
}
\arguments{
\item{object}{a \code{clust_analysis} or \code{combi_analysis} object.}

\item{...}{extra arguments passed to \code{\link{np}}.}
}
\value{
a data frame with columns characterized in Details.
}
\description{
Computes basic global statistics of quality of a clustering analysis object.
}
\details{
The statistics retrieved by the \code{summary()} method are:
\itemize{
\item \emph{silhouette width} (\code{sil_width})
\item \emph{fraction of potentially misclassified observations} with negative
silhouette widths (\code{frac_misclassified})
\item \emph{fraction of explained clustering variance} expressed as the ratio of total
between sum of squares to total sum of squares (\code{frac_var})
\item \emph{fraction of preserved nearest neighbors} (\code{frac_np})
}

The statistics are computed with \code{\link{silhouette}}, \code{\link{var}},
and \code{\link{np}} methods for the entire clustering structure and not
for particular clusters.
}
\references{
Rousseeuw PJ. Silhouettes: A graphical aid to the interpretation and
validation of cluster analysis. J Comput Appl Math (1987) 20:53–65.
doi:10.1016/0377-0427(87)90125-7

Venna J, Kaski S. Neighborhood preservation in nonlinear projection methods:
An experimental study. Lect Notes Comput Sci (including Subser Lect Notes
Artif Intell Lect Notes Bioinformatics) (2001) 2130:485–491.
doi:10.1007/3-540-44668-0_68
}
