% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_dist_oop.R
\name{summary.cross_dist}
\alias{summary.cross_dist}
\title{Summary of cross-distances between clusters.}
\usage{
\method{summary}{cross_dist}(object, ...)
}
\arguments{
\item{object}{a \code{cross_dist} object.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with mean, SD, median, interquartile range, 95\% range
and range of cross-distances between the cluster pairs.
}
\description{
Computes summary statistics of homologous or heterologous cross-distances
between the clusters.
}
