% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_oop.R
\name{summary.importance}
\alias{summary.importance}
\title{Importance statistic summary.}
\usage{
\method{summary}{importance}(object, ...)
}
\arguments{
\item{object}{an `importance` class object.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with importance metrics.
}
\description{
If the permutation importance analysis for clustering variables was done
in multiple iterations
(e.g. `n_iter` set to > 1 in \code{\link{impact.clust_analysis}}),
number of iterations, mean, SD, median, interquartile range
and range of the difference in
clustering variance for each clustering variable is computed.
Otherwise, a plain data frame with importance statistics is returned.
}
