% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knb_oop.R
\name{summary.knb}
\alias{summary.knb}
\title{Summary neighborhood preservation statistic for clusters.}
\usage{
\method{summary}{knb}(object, ...)
}
\arguments{
\item{object}{an object of the \code{\link{knb}} class.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with numeric statistics for the whole clustering
structure (\code{clust_id} = 'global') and particular clusters.
}
\description{
Computes mean, SD, median, interquartile range, 95\% range and range of the
network preservation statistic for the global clustering structure
and particular clusters.
Low values of the median or median neighborhood preservation statistic may
indicate that the given cluster s poorly separated from other clusters.
For objects without cluster assignment (e.g.
neighborhood analyses for reduction analysis methods), only global
neighborhood statistics are returned.
For such objects, low values of the neighborhood preservation statistic
suggest that the reduction analysis poorly projects the data point
neighborhood into the reduced layout.
}
\references{
Venna J, Kaski S. Neighborhood preservation in nonlinear projection methods:
An experimental study. Lect Notes Comput Sci (including Subser Lect Notes
Artif Intell Lect Notes Bioinformatics) (2001) 2130:485–491.
doi:10.1007/3-540-44668-0_68#'

Breard GT. Evaluating Self-Organizing Map Quality Measures as Convergence
Criteria Criteria. Open Access Master’s Theses. Paper 1033.
Available at: https://digitalcommons.uri.edu/theses/1033
}
