% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_analysis_oop.R
\name{summary.red_analysis}
\alias{summary.red_analysis}
\title{Summary of a red_analysis object.}
\usage{
\method{summary}{red_analysis}(object, ...)
}
\arguments{
\item{object}{a red_analysis object.}

\item{...}{extra arguments, currently none.}
}
\description{
Summary method for the red_analysis class. Prints the summary
of the reduction analysis object.
}
