% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sil_oop.R
\name{summary.sil_extra}
\alias{summary.sil_extra}
\title{Summary silhouette width statistic for clusters.}
\usage{
\method{summary}{sil_extra}(object, ...)
}
\arguments{
\item{object}{an object of the \code{\link{sil_extra}} class.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with numeric statistics for the whole clustering
structure (`clust_id` = 'global') and particular clusters.
}
\description{
Computes mean, SD, median, interquartile range, 95% range, range as well as
the number and percentage of observations with negative silhouette width.
Such observations are likely in an improper cluster.
}
\references{
Rousseeuw PJ. Silhouettes: A graphical aid to the interpretation and
validation of cluster analysis. J Comput Appl Math (1987) 20:53–65.
doi:10.1016/0377-0427(87)90125-7
}
