% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuner_oop.R
\name{summary.tuner}
\alias{summary.tuner}
\title{Summary of quality statistics.}
\usage{
\method{summary}{tuner}(object, ...)
}
\arguments{
\item{object}{a \code{tuner} object.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with the following columns:
\itemize{
\item \code{sil_width}: silhouette width
\item \code{frac_misclassified}: fraction of observations with negative silhouette
widths suggestive of misclassification
\item \code{frac_var}: fraction of explained clustering variance
\item \code{frac_np}: fraction of preserved nearest neighbors
\item columns named after names of the tuning parameters and containing their
values
}
}
\description{
The \code{summary()} method called for \code{tuner} class objects extracts cluster
assignment prediction statistics for subsequent combinations of the tuning
parameters from the object.
}
