% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.R
\name{te}
\alias{te}
\alias{te.clust_analysis}
\alias{te.combi_analysis}
\title{Topology error of self-organizing maps.}
\usage{
te(x, ...)

\method{te}{clust_analysis}(x, ...)

\method{te}{combi_analysis}(x, type = c("node", "final"), ...)
}
\arguments{
\item{x}{a \code{clust_analysis} or \code{combi_analysis} object.}

\item{...}{extra arguments passed to methods.}

\item{type}{type reference clusters. For \code{type = 'node'}, topology error
within SOM nodes is computed, for \code{type = 'final'}, topology error in the
final clusters (clusters of SOM nodes) is calculated.}
}
\value{
An object of the \code{\link{knb}} class with
\code{\link{summary.knb}} and \code{\link{plot.knb}} methods.

An object of the \code{\link{knb}} class with
\code{\link{summary.knb}} and \code{\link{plot.knb}} methods.
}
\description{
Calculates the topology error for data points in a self-organizing map (SOM).
}
\details{
The procedure of topology error computation is as follows: for each
observation, two nearest self-organizing map (SOM) nodes are identified.
If such nodes are neighbors in the initial layout of SOM prior to data
fitting, correct topology (coded with 0) is returned and an error otherwise
(coded as 1).
The function returns NULL with a warning when called for a non-SOM analysis
object.
\code{te()} is a S3 generic function.
}
