% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{tuner}
\alias{tuner}
\title{Create an \code{tuner} class object.}
\usage{
tuner(x)
}
\arguments{
\item{x}{a list with elements specified in Details.}
}
\value{
an instance of the \code{tuner} class as described in Details.
}
\description{
Creates and object of the \code{tuner} class on the top of a list with tuning
of parameters of cluster analysis or prediction.
}
\details{
The input list has to have three elements:
\itemize{
\item \code{analysis}: a \code{clust_analysis} or \code{combi_analysis} object created with the
best set of the tuning parameters
\item \code{stats}: a data frame with values of quality stats (silhouette width,
fraction of potentially misclassified observations/negative silhouette width,
fraction of explained clustering variance, and fraction of preserved nearest
neighbors)
\item \code{fun}: name of the tuning function
\item \code{dataset}: a string specified which data was used during the tuning: the
training data set or cross-validation
\item \code{type}: type of analysis, development or prediction
\item \code{clust_vars}: a vector of names of clustering variables
\item \code{tune_params}: a vector of names of the tuning parameters
\item \code{tune_criteria}: a data frame that specifies which criteria were applied
to select the best combination of the tuning parameters
\item \code{best_tune}: a data frame storing the best values of the tuning parameters
}
}
