% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_functions.R
\name{umap}
\alias{umap}
\title{Uniform Manifold Approximation and Projection (UMAP).}
\usage{
umap(data, distance_method, kdim, ...)
}
\arguments{
\item{data}{a data frame.}

\item{distance_method}{name of the distance metric specified by a
\code{\link[umap]{umap.defaults}} object.}

\item{kdim}{dimension number.}

\item{...}{extra arguments passed to \code{\link[umap]{umap}}.}
}
\description{
Performs UMAP of the data. A wrapper
around \code{\link[umap]{umap}}.
}
\details{
UMAP parameters such as dimension number or distance are provided
as a \code{\link[umap]{umap.defaults}} object.
}
