% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{var.clust_analysis}
\alias{var.clust_analysis}
\alias{var.combi_analysis}
\title{Calculate clustering variance.}
\usage{
\method{var}{clust_analysis}(x, ...)

\method{var}{combi_analysis}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments, currently none.}
}
\value{
a list with the total, within-cluster, between-cluster sum of
squares and explained clustering variance (named `frac_var`).
}
\description{
Calculates the clustering sum of squares (total, within
clusters, total within clusters and between clusters) as well as the
fraction of 'explained' clustering variance. The later is the ratio of
the total between-cluster sum of squares to the total sum of squares.
}
\details{
`var()` is a S3 generic function.
`var()` overwrites the `var()` function provided by the stats package,
but provides a handy default method, so that `var()` is expected to behave
the same way as in base R.
}
