% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_analysis_oop.R
\name{var.combi_analysis}
\alias{var.combi_analysis}
\title{Calculate clustering variance.}
\usage{
\method{var}{combi_analysis}(x, ...)
}
\arguments{
\item{x}{a combi_analysis object.}

\item{...}{extra arguments, currently none.}
}
\description{
Calculates the clustering sum of squares (total, within
clusters, total within clusters and between clusters) as well as the
fraction of 'explained' clustering variance. The later is the ratio of
between-cluster sum of squares to the total sum of squares statistic.
}
