% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{var.default}
\alias{var.default}
\title{Calculate object's variance.}
\usage{
\method{var}{default}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to \code{\link[stats]{var}}.}
}
\description{
Default variance calculation method, a wrapper around
\code{\link[stats]{var}}.
}
