% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{var}
\alias{var}
\title{Calculate object's variance.}
\usage{
var(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\description{
Calculates variance of an object.
}
