% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_analysis_oop.R
\name{var.red_analysis}
\alias{var.red_analysis}
\title{component variance.}
\usage{
\method{var}{red_analysis}(x, ...)
}
\arguments{
\item{x}{a red_analysis object.}

\item{...}{extra arguments, currently none.}
}
\description{
Computes the standard variations and variances associated with
the dimensions/component of a red_analysis object.
}
