% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vote_kernel}
\alias{vote_kernel}
\title{Find the most frequently occurring element with distance weighting.}
\usage{
vote_kernel(vector, dist_vec, kernel_fun = function(x) 1/x)
}
\arguments{
\item{vector}{a vector.}

\item{dist_vec}{a numeric vector with the distance values.}

\item{kernel_fun}{a kernel function.}
}
\description{
Finds the element of a vector with the highers number of
occurrences. The voting is distance weighted by the given kernel function.
}
\references{
the most frequent element.
}
