% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vote_simple}
\alias{vote_simple}
\title{Find the most frequently occurring element of a vector.}
\usage{
vote_simple(vector, resolve_ties = FALSE)
}
\arguments{
\item{vector}{a vector.}

\item{resolve_ties}{logical, should the ties be resolved at random?}
}
\value{
the most frequent element.
}
\description{
Finds the element of a vector with the highers number of
occurrences.
}
\details{
Ties may be resolved at random (resolve_ties = TRUE), otherwise,
if a tie exists, the alphabetically first element is returned.
}
