% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_dist}
\alias{calculate_dist}
\title{Calculate distance.}
\usage{
calculate_dist(data, method)
}
\arguments{
\item{data}{a data frame or matrix. Rownames are preserved as
observation IDs.}

\item{method}{the name of dissimilarity measure, see:
\code{\link{get_kernel_info}} for available distances.}
}
\value{
a matrix with the distance statistics.
}
\description{
Calculates distances between observations (rows) of a data
frame or matrix. See \code{\link{get_kernel_info}} for a vector of available
distance measures.
}
\details{
Provides a handy wrapper for \code{\link[philentropy]{distance}}.
The smc (simple matching coefficient) distance is calculated with the
\code{\link[nomclust]{sm}} function.
}
