% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_dist}
\alias{calculate_dist}
\alias{get_kernel_info}
\title{Calculate distance.}
\usage{
calculate_dist(data, method)

get_kernel_info()
}
\arguments{
\item{data}{a data frame or matrix. Row names are preserved as
observation IDs.}

\item{method}{the name of dissimilarity measure, see:
\code{\link{get_kernel_info}} for available distances.}
}
\value{
`calculate_dist()`: a matrix with the distance statistics;
`get_kernel_info()`: a vector with names of available distance measures.
}
\description{
The `calculate_dist()` function calculates distances
between observations (rows) of a data
frame or matrix. See \code{\link{get_kernel_info}} for a vector of available
distance measures.
}
\details{
Provides a handy wrapper for \code{\link[philentropy]{distance}}.
The smc (simple matching coefficient) distance is calculated with the
\code{\link[nomclust]{sm}} function.
}
\references{
Drost H-G. Philentropy: Information Theory and Distance Quantification
with R. J Open Source Softw (2018) 3:765. doi:10.21105/joss.00765

Boriah S, Chandola V, Kumar V. Similarity measures for categorical data:
A comparative evaluation. in Society for Industrial and
Applied Mathematics - 8th SIAM International Conference on Data Mining
2008, Proceedings in Applied Mathematics 130, 243–254.
doi:10.1137/1.9781611972788.22

Sulc Z, Cibulkova J, Rezankova H. nomclust: Hierarchical Cluster Analysis
of Nominal Data. (2021)
Available at: https://cran.r-project.org/package=nomclust
}
