% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{center_data}
\alias{center_data}
\alias{min_max}
\title{Normalization of a data frame.}
\usage{
center_data(data, type = c("mean", "median"), complete_cases = FALSE)

min_max(data, complete_cases = FALSE)
}
\arguments{
\item{data}{a data frame or a tibble. All variables need to be numeric.}

\item{type}{type of the centering, mean (default) or median.}

\item{complete_cases}{logical, should the observations with the complete
variable record only be included as an output?}
}
\value{
a data frame or a tibble.
}
\description{
Normalization with median or mean centering of a data frame
or tibble (\code{center_data()}) or simple min/max normalization (\code{min_max()}).
Preserves the row names.
}
\details{
A wrapper around \code{\link[base]{scale}}. Mean scaling is equal
to canonical Z-score normalization.
}
