% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{clust_analysis}
\alias{clust_analysis}
\title{Construct a clust_analysis class object.}
\usage{
clust_analysis(x)
}
\arguments{
\item{x}{a named list, see details.}
}
\description{
Constructs a clust_analysis class object given a list storing
results of a clustering analysis.
}
\details{
A named list with the following elements is required:
data (a quosure calling the original data set),
dist_mtx (a matrix with the distances between the observations),
dist_method (name of the distance statistic),
clust_obj (the output object of the clustering analysis),
clust_fun (the name of the clustering function),
clust_assignment (a tibble with the cluster assignment of the
observations with the observation and clust_id variables),
dots (additional arguments passed to the clustering function).
}
