% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{combi_analysis}
\alias{combi_analysis}
\title{Combined SOM - clustering analysis.}
\usage{
combi_analysis(x)
}
\arguments{
\item{x}{a named list, see Details.}
}
\value{
a \code{combi_analysis} object with the elements specified in Details.
}
\description{
Constructs a \code{combi_anlysis()} class object given a list with
results of reduction analysis or self-organizing map and clustering analysis.
}
\details{
A named list with the following elements is required as the \code{x} argument:
\itemize{
\item \code{clust_analyses}: a list of \code{red_analysis} or \code{clust_analysis} objects,
\item \code{clust_assignment}: a data frame with the cluster assignment with the
\code{observation} and \code{clust_id} variables.
}

For combined solutions involving unsupervised clustering of the SOM U matrix,
the function returns an object of subclass
\code{umatrix_analysis}, which inherits virtually all methods from the
\code{combi_analysis} class.
}
