% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_clustering.R
\name{combi_cluster}
\alias{combi_cluster}
\title{Cluster self-organizing map nodes.}
\usage{
combi_cluster(
  data,
  distance_som = "euclidean",
  xdim = 5,
  ydim = 4,
  topo = "hexagonal",
  neighbourhood.fct = "gaussian",
  toroidal = FALSE,
  rlen = 500,
  node_clust_fun = hcluster,
  distance_nodes = "euclidean",
  seed = 1234,
  ...
)
}
\arguments{
\item{data}{a numeric data frame, matrix or a red_analysis object.
If a red_analysis object is provided, its component/score table will be
clustered.}

\item{distance_som}{metric of distance between the observations, used for SOM
development. See: \code{\link{get_kernel_info}}.}

\item{xdim}{x dimension of the SOM grid,
see: \code{\link[kohonen]{somgrid}} for details.}

\item{ydim}{y dimension of the SOM grid,
#' see: \code{\link[kohonen]{somgrid}} for details.}

\item{topo}{SOM grid topology, see: \code{\link[kohonen]{somgrid}}
for details. 'hexagonal' for default.}

\item{neighbourhood.fct}{neighborhood function, 'gaussian' for default.}

\item{toroidal}{logical, should toroidal grid be used?}

\item{rlen}{number of the SOM algorithm iterations.}

\item{node_clust_fun}{a function provided by the clustTools package used to
cluster the SOM nodes. Alternatively, a custom function may be provided here,
which returns a clust_analysis class object.}

\item{distance_nodes}{metric of distance between the nodes, used for SOM
development. See: \code{\link{get_kernel_info}}.}

\item{seed}{initial setting of the random number generator.}

\item{...}{extra arguments passed to node_clust_fun, such as k number of
clusters.}
}
\description{
Performs clustering of the self-orgnizing map (SOM) nodes with
one of the clustering functions provided by the clustTools package.
}
