% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis_oop.R
\name{components.clust_analysis}
\alias{components.clust_analysis}
\title{Dimensionality reduction analysis of the clust_analysis object data
or distance matrix.}
\usage{
\method{components}{clust_analysis}(
  x,
  kdim = NULL,
  red_fun = c("pca", "mds", "umap"),
  with = c("distance", "data"),
  ...
)
}
\arguments{
\item{x}{a clust_analysis object.}

\item{kdim}{number of dimensions. If NULL, kdim is set to the number of
clusters.}

\item{red_fun}{reduction analysis function: 'pca' (PCA), 'mds' (MDS) or
'umap' (UMAP).}

\item{with}{type of the input data for the reduction analysis:
the clustering data ('data') or the matrix of distances ('distance').}

\item{...}{extra arguments passed to \code{\link{reduce_data}}.}
}
\value{
a red_analysis object with the component/score table containing the
cluster assignment information ('clust_id' variable).
}
\description{
Performs principal component analysis (PCA), multi-dimensional
scaling (MDS) or uniform manifold approximation and projection (UMAP) of the
clust_analysis object data or distance matrix.
}
\details{
See \code{\link{reduce_data}} for the implementation details.
The distance method, relevant for MDS and UMAP. is taken over from the
clust_object. Hence, some distances may crash the analysis with UMAP, see:
\code{\link[umap]{umap.defaults}} for the compatible distances.
}
