% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{components.clust_analysis}
\alias{components.clust_analysis}
\alias{components.min_analysis}
\alias{components.combi_analysis}
\alias{components.umatrix_analysis}
\title{Dimensionality reduction analysis of the analysis data or distance matrix.}
\usage{
\method{components}{clust_analysis}(
  object,
  kdim = NULL,
  red_fun = c("pca", "mds", "umap"),
  with = c("distance", "data", "umatrix"),
  train_object = NULL,
  ...
)

\method{components}{min_analysis}(object, kdim = NULL, red_fun = c("pca", "mds", "umap"), ...)

\method{components}{combi_analysis}(
  object,
  kdim = NULL,
  red_fun = c("pca", "mds", "umap"),
  with = c("distance", "data", "umatrix"),
  train_object = NULL,
  ...
)

\method{components}{umatrix_analysis}(
  object,
  kdim = NULL,
  red_fun = c("pca", "mds", "umap"),
  with = c("distance", "data", "umatrix"),
  ...
)
}
\arguments{
\item{object}{an object.}

\item{kdim}{number of dimensions. If NULL, kdim is set to the number of
clusters.}

\item{red_fun}{reduction analysis function: 'pca' (PCA), 'mds' (MDS) or
'umap' (UMAP).}

\item{with}{type of the input data for the reduction analysis:
the clustering data ('data'), the matrix of distances between observations
('distance') or U matrix between SOM nodes ('umatrix').}

\item{train_object}{an optional \code{red_analysis} object wchich will be used as
a training layout. This works currently only for PCA and UMAP and in case
of single layer data frames as clustering data.
Please refer to \code{\link[stats]{prcomp}} and \code{\link[umap]{umap}} for
details.}

\item{...}{extra arguments passed to \code{\link{reduce_data}}.}
}
\value{
a \code{red_analysis} object with the component/score table containing
the cluster assignment information (\code{clust_id} variable). In case of
multi-layer SOM analyses with the clustering data set, a list of
\code{red_analysis} objects is returned.
}
\description{
Performs principal component analysis (PCA), multi-dimensional
scaling (MDS) or uniform manifold approximation and projection (UMAP) of the
analysis data set used for clustering or distance matrix.
}
\details{
See \code{\link{reduce_data}} for the implementation details.
The distance method, relevant for MDS and UMAP. is taken over from the
\code{clust_object}. Hence, some distances may crash the analysis with UMAP, see:
\code{\link[umap]{umap.defaults}} for the compatible distances.
For \code{combi_analysis} objects, the analysis is done for the global clustering,
i.e. assignment of observations to the clusters and not to the SOM nodes.
In cases, when the clustering analysis was done with an user-provided
dissimilarity object (subclass \code{min_analysis} of \code{clust_analysis}) it is not
possible to perform a dimensionality reduction analysis with the genuine data
set - the analysis can be performed only for the distance matrix.
In case of multi-layer SOM analyses or combined multi-layer SOM - clustering
analyses and \code{with} set to 'data', the dimensionality reduction analysis
is done separately for each data layer.
}
