% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cv}
\alias{cv}
\alias{cv.clust_analysis}
\alias{cv.combi_analysis}
\title{Cross-validate the clustering analysis object.}
\usage{
cv(x, ...)

\method{cv}{clust_analysis}(
  x,
  nfolds = 5,
  kNN = 5,
  simple_vote = TRUE,
  resolve_ties = FALSE,
  kernel_fun = function(x) 1/x,
  seed = 1234,
  .parallel = FALSE,
  ...
)

\method{cv}{combi_analysis}(
  x,
  nfolds = 5,
  kNN = 5,
  simple_vote = TRUE,
  resolve_ties = FALSE,
  kernel_fun = function(x) 1/x,
  seed = 1234,
  .parallel = FALSE,
  ...
)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments, currently none.}

\item{nfolds}{number of CV folds.}

\item{kNN}{number of the nearest neighbors.}

\item{simple_vote}{logical, should classical unweighted k-NN classification
be applied? If FALSE, distance-weighted k-NN is used with the provided kernel
function.}

\item{resolve_ties}{logical, should the ties be resolved at random? Applies
only to the simple unweighted voting algorithm.}

\item{kernel_fun}{kernel function transforming the distance into weight.}

\item{seed}{initial setting of the random number generator.}

\item{.parallel}{logical, should the CV be run in parallel?}
}
\value{
a list containing the global \code{\link{clust_analysis}} object,
projection (prediction) results and prediction summary
for each fold and a prediction
summary for the whole CV.
}
\description{
Checks the stability of a clustering solution by
cross-validation (CV) and the classification error in CV folds
as a measure of the cluster stability.
}
\details{
By principle similar to cross-validation of any machine learning
multi-class classifier. The training portion of a CV split is used to develop
of a cluster structure and the projection on the test portion is accomplished
by k-nearest neighbor (kNN) label propagation algorithm. For its
implementation details, see: \code{\link{propagate}}.
For `combi_analysis` objects, assignment of the observations in the CV folds
is done for the 'top' assignment of the observations to the clusters;
nodes are ignored!
The fold are generated with \code{\link[caret]{createFolds}}.
`cv()` is a S3 generic function.
}
\references{
Lange T, Roth V, Braun ML, Buhmann JM. Stability-based validation of
clustering solutions. Neural Comput (2004) 16:1299–1323.
doi:10.1162/089976604773717621

Leng M, Wang J, Cheng J, Zhou H, Chen X. Adaptive semi-supervised
clustering algorithm with label propagation. J Softw Eng (2014) 8:14–22.
doi:10.3923/jse.2014.14.22

Kuhn M. Building predictive models in R using the caret package.
J Stat Softw (2008) 28:1–26. doi:10.18637/jss.v028.i05
}
