% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dist}
\alias{dist}
\title{Distance Matrix Computation.}
\usage{
dist(x, ...)
}
\arguments{
\item{x}{a numeric matrix, data frame or 'dist' object.}

\item{...}{extra arguments passed to methods.}
}
\description{
Calculates or extracts a distance matrix.
}
