% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dist}
\alias{dist}
\alias{dist.default}
\title{Distance between observations.}
\usage{
dist(x, ...)

\method{dist}{default}(x, ...)
}
\arguments{
\item{x}{an object. For the default method a numeric matrix, data frame
or \code{dist} object.}

\item{...}{arguments for methods, e.g. passed to \code{\link[stats]{dist}}.}
}
\description{
Computes the distance between observations in a matrix, data frame
or other compatible objects.
}
\details{
The default \code{dist()} method is a wrapper around
\code{\link[stats]{dist}}.
}
