% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extract}
\alias{extract}
\alias{extract.clust_analysis}
\alias{model.frame.clust_analysis}
\alias{dist.clust_analysis}
\alias{extract.combi_analysis}
\alias{model.frame.combi_analysis}
\alias{dist.combi_analysis}
\alias{extract.red_analysis}
\alias{model.frame.red_analysis}
\title{Extract features of a clust_analysis object.}
\usage{
extract(x, ...)

\method{extract}{clust_analysis}(
  x,
  type = c("distance", "assignment", "clust_object", "data", "object"),
  ...
)

\method{model.frame}{clust_analysis}(formula, ...)

\method{dist}{clust_analysis}(x, ...)

\method{extract}{combi_analysis}(
  x,
  type = c("distance", "assignment", "clust_object", "data", "object"),
  ...
)

\method{model.frame}{combi_analysis}(formula, ...)

\method{dist}{combi_analysis}(x, ...)

\method{extract}{red_analysis}(
  x,
  type = c("component_tbl", "scores", "loadings", "data", "sdev", "object"),
  ...
)

\method{model.frame}{red_analysis}(formula, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments, currently none.}

\item{type}{the feature name:
`distance` extracts the matrix with distances between the observations,
`data` the data set used for the analysis,
`assignment` assignment of the observations to the clusters,
`clust_object` or `object` returns the wrapped clustering object.
`clust_object` or `scores` return the component pr score tables for the
observations,
`loadings` retrieves the table of variable loadings (PCA only),
`sdev` returns standard deviations, associated with the
components (PCA only).}

\item{formula}{an object.}
}
\value{
the requested feature/property.
}
\description{
A general extractor method for accessing properties and features
of a `clust_analysis`, `combi_analysis` and `red_analysis` object,
and specific methods for accessing the modeling data frame
and distance matix.
}
\details{
`extract()` is a S3 generic function.
}
