% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{hcluster}
\alias{hcluster}
\title{Hierarchical clustering.}
\usage{
hcluster(
  data,
  distance_method = "euclidean",
  k = 2,
  hc_method = "ward.D2",
  seed = 1234,
  ...
)
}
\arguments{
\item{data}{a numeric data frame or matrix or a red_analysis object.}

\item{distance_method}{name of the distance metric, see:
\code{\link{get_kernel_info}}.}

\item{k}{number of clusters.}

\item{hc_method}{the hierarchical clustering algorithm, see:
\code{\link[stats]{hclust}} for details.}

\item{seed}{initial setting of the random number generator.}

\item{...}{extra arguments passed to \code{\link[stats]{hclust}}.}
}
\value{
a clust_analysis object.
}
\description{
Performs hierarchical clustering analysis of a numeric data
frame, matrix or the results of a reduction analysis.
}
\details{
Technically, a wrapper around \code{\link[stats]{hclust}}. If a
red_analysis object is provided as the data argument, the observation
component/score table is subjected to clustering.
}
