% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{hcluster}
\alias{hcluster}
\title{Hierarchical clustering.}
\usage{
hcluster(
  data,
  distance_method = "euclidean",
  k = 2,
  hc_method = "ward.D2",
  seed = 1234,
  ...
)
}
\arguments{
\item{data}{a numeric data frame or matrix or a \code{red_analysis} object.}

\item{distance_method}{name of the distance metric, see:
\code{\link{get_kernel_info}}. Ignored if data is provided as
a distance matrix.}

\item{k}{number of clusters.}

\item{hc_method}{the hierarchical clustering algorithm, see:
\code{\link[stats]{hclust}} for details.}

\item{seed}{initial setting of the random number generator.}

\item{...}{extra arguments passed to \code{\link[stats]{hclust}}.}
}
\value{
an object of the class \code{\link{clust_analysis}}.
}
\description{
Performs hierarchical clustering analysis of a numeric data
frame, matrix or the results of a reduction analysis.
}
\details{
Technically, a wrapper around \code{\link[stats]{hclust}}. If a
\code{red_analysis} object is provided as the \code{data} argument, the observation
component/score table is subjected to clustering.
If a dissimilarity object of class \code{dist} is provided as \code{data}, the function
returns an instance of subclass of the \code{\link{clust_analysis}} class,
the \code{min_analysis} object. Such objects inherit most of the methods of the
fully fledged clustering analysis object. Yet, methods requiring the source
data frame such as reduction analysis with the genuine data, variable
importance analysis or computation of heterologous cross-distances
are not possible. Semi-supervised clustering and cross-validation are not
implemented at the moment for the
\code{min_analysis} objects.
}
\references{
Murtagh F, Contreras P. Algorithms for hierarchical clustering:
An overview. Wiley Interdiscip Rev Data Min Knowl Discov
(2012) 2:86–97. doi:10.1002/widm.53
}
