% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{impact}
\alias{impact}
\title{Get variable importance statistics.}
\usage{
impact(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods.}
}
\description{
Computes variable importance statistics.
}
