% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{importance}
\alias{importance}
\title{Construct an importance object.}
\usage{
importance(x)
}
\arguments{
\item{x}{a data frame with the columns 'total_wss', 'total_ss', 'between_ss',
'frac_var', 'variable' and 'frac_diff'.}
}
\value{
a tibble of the `importance` class.
}
\description{
Constructs an object of class `importance` on the top of
a tibble with the importance testing results.
}
