% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_testing.R
\name{is_clust_analysis}
\alias{is_clust_analysis}
\alias{is_combi_analysis}
\alias{is_red_analysis}
\alias{is_importance}
\alias{is_cross_dist}
\alias{is_sil_extra}
\title{Test class inheritance.}
\usage{
is_clust_analysis(x)

is_combi_analysis(x)

is_red_analysis(x)

is_importance(x)

is_cross_dist(x)

is_sil_extra(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests if the object is an instance of the `red_analysis`, `clust_analysis`,
`combi_analysis`, `importance`, `cross_dist` or `sil_extra` class.
}
