% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_testing.R
\name{is_clust_analysis}
\alias{is_clust_analysis}
\alias{is_min_analysis}
\alias{is_umatrix_analysis}
\alias{is_combi_analysis}
\alias{is_red_analysis}
\alias{is_importance}
\alias{is_cross_dist}
\alias{is_sil_extra}
\alias{is_cluster_cv}
\title{Test class inheritance.}
\usage{
is_clust_analysis(x)

is_min_analysis(x)

is_umatrix_analysis(x)

is_combi_analysis(x)

is_red_analysis(x)

is_importance(x)

is_cross_dist(x)

is_sil_extra(x)

is_cluster_cv(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests if the object is an instance of the \code{red_analysis}, \code{clust_analysis},
\code{combi_analysis}, \code{importance}, \code{cross_dist}, \code{sil_extra}, \code{min_analysis},
\code{umatrix_analysis} or \code{clusetr_cv} class.
}
