% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{kcluster}
\alias{kcluster}
\title{K-means or medoid clustering.}
\usage{
kcluster(
  data,
  distance_method = "euclidean",
  clust_fun = c("kmeans", "pam"),
  k = 2,
  seed = 1234,
  ...
)
}
\arguments{
\item{data}{a numeric data frame or matrix or a red_analysis object.}

\item{distance_method}{name of the distance metric, see:
\code{\link{get_kernel_info}}.}

\item{clust_fun}{the name of the clustering function, currently implemented
are 'kmeans' and 'pam'.}

\item{k}{number of clusters.}

\item{seed}{initial setting of the random number generator.}

\item{...}{extra arguments passed to \code{\link[stats]{kmeans}} or
\code{\link[cluster]{pam}}.}
}
\value{
an object of the class \code{\link{clust_analysis}}.
}
\description{
Performs k-means and PAM (partition around medoids) clustering
analysis of of a numeric data frame, matrix or the results of
a reduction analysis.
}
\details{
Technically, a wrapper around \code{\link[stats]{kmeans}} and
\code{\link[cluster]{pam}}. If a red_analysis object is provided as the
data argument, the observation component/score table is subjected to
clustering.
}
\references{
Hartigan JA, Wong MA. Algorithm AS 136: A K-Means Clustering Algorithm.
Appl Stat (1979) 28:100. doi:10.2307/2346830

Schubert E, Rousseeuw PJ. Faster k-Medoids Clustering: Improving the PAM,
CLARA, and CLARANS Algorithms. in Lecture Notes in Computer Science
(including subseries Lecture Notes in Artificial Intelligence and
Lecture Notes in Bioinformatics) (Springer), 171–187.
doi:10.1007/978-3-030-32047-8_16
}
