% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_functions.R
\name{mds}
\alias{mds}
\alias{pca}
\alias{umap}
\alias{fa}
\title{Perform dimensionality reduction.}
\usage{
mds(data, distance_method = "euclidean", kdim = 2, ...)

pca(data, kdim = 2, ...)

umap(data, distance_method, kdim, ...)

fa(data, kdim = 2, ...)
}
\arguments{
\item{data}{a data frame.}

\item{distance_method}{name of the distance metric, see:
\code{\link{get_kernel_info}}}

\item{kdim}{dimension number.}

\item{...}{extra arguments passed to
\code{\link[stats]{cmdscale}}, \code{\link[pcaPP]{PCAproj}},
\code{\link[umap]{umap}} or \code{\link[stats]{factanal}}.}
}
\value{
an object of the class \code{\link{red_analysis}}.
}
\description{
Performs
multi-dimensional scaling (`mds()` via ' \code{\link[stats]{cmdscale}}),
principal component analysis (`pca()` via \code{\link[pcaPP]{PCAproj}}),
UMAP (`umap()` via \code{\link[umap]{umap}}),
or factor analysis (`fa()` via \code{\link[stats]{factanal}}).
}
\details{
UMAP parameters such as dimension number or distance are provided
as a \code{\link[umap]{umap.defaults}} object.
}
\references{
McInnes L, Healy J, Melville J. UMAP: Uniform Manifold Approximation and
Projection for Dimension Reduction. (2018)
Available at: https://arxiv.org/abs/1802.03426v3

Croux C, Filzmoser P, Oliveira MR. Algorithms for Projection-Pursuit robust
principal component analysis. Chemom Intell Lab Syst (2007) 87:218–225.
doi:10.1016/j.chemolab.2007.01.004

BARTLETT MS. THE STATISTICAL CONCEPTION OF MENTAL FACTORS. Br J Psychol
Gen Sect (1937) 28:97–104. doi:10.1111/j.2044-8295.1937.tb00863.x
}
