% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{ngroups}
\alias{ngroups}
\alias{ngroups.clust_analysis}
\alias{ngroups.combi_analysis}
\title{Numbers of observations in the clusters.}
\usage{
ngroups(x, ...)

\method{ngroups}{clust_analysis}(x, ...)

\method{ngroups}{combi_analysis}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{extra arguments passed to methods, currently none.}
}
\description{
Compute numbers of observations in the clusters or, for `combi_analysis`
objects, numbers of observations in the SOM nodes and clusters.
}
\details{
`ngroups()` is a S3 generic function.
}
