% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{nobs.clust_analysis}
\alias{nobs.clust_analysis}
\alias{nobs.red_analysis}
\alias{nobs.combi_analysis}
\title{Number of observations and variables for dimensionality reduction and clustering.}
\usage{
\method{nobs}{clust_analysis}(object, ...)

\method{nobs}{red_analysis}(object, ...)

\method{nobs}{combi_analysis}(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{extra arguments, currently none.}
}
\value{
a list with the numbers of observations and variables.
}
\description{
Computes numbers of observations and variables used in the analyses.
}
