% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis_oop.R
\name{nobs.clust_analysis}
\alias{nobs.clust_analysis}
\title{Number of observations of a clust_analysis object.}
\usage{
\method{nobs}{clust_analysis}(object, ...)
}
\arguments{
\item{object}{a clust_analysis object.}

\item{...}{extra arguments, currently none.}
}
\value{
a list with the numbers of observations and clustering features.
}
\description{
Retrieves the number of observations used to generate
a clust_analysis object.
}
