% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combi_analysis_oop.R
\name{plot.combi_analysis}
\alias{plot.combi_analysis}
\alias{plot.umatrix_analysis}
\title{Plot selected features of a combi_analysis object.}
\usage{
\method{plot}{combi_analysis}(
  x,
  type = c("diagnostic", "components", "heat_map", "training", "data"),
  cust_theme = ggplot2::theme_classic(),
  jitter_width = 0,
  jitter_height = 0,
  point_alpha = 1,
  ...
)

\method{plot}{umatrix_analysis}(
  x,
  type = c("diagnostic", "components", "heat_map", "training", "data"),
  cust_theme = ggplot2::theme_classic(),
  jitter_width = 0,
  jitter_height = 0,
  point_alpha = 1,
  ...
)
}
\arguments{
\item{x}{a \code{combi_analysis} object.}

\item{type}{the type of plots:
\itemize{
\item \code{diagnostic} returns a series of diagnostic plots for the SOM construction
and clustering of the nodes
\item \code{components} plots the results of reduction analysis done with the
clustering data or the distance matrix
(see: \code{\link{components.clust_analysis}})
\item \code{heat_map} plots the distances between observations and nodes as heat maps
\item \code{training} plots the mean distance to the SOM winning unit as a function
of the iteration number
\item \code{data} works only if reduction analysis results were used for clustering
and plots the first two components/dimensions.
}}

\item{cust_theme}{a ggplot theme.}

\item{jitter_width}{horizontal jittering of the points in the plots.}

\item{jitter_height}{vertical jittering of the points in the plots.}

\item{point_alpha}{scatter plot's point alpha.}

\item{...}{extra arguments passed to \code{\link{components.clust_analysis}}.}
}
\value{
a ggplot object or a list of ggplot objects, as specified by the
'type' argument
}
\description{
The plotting method for the \code{combi_analysis} class. Enables
plotting of the standard diagnostic plots used for the optimal cluster number
determination for the node clustering (dendrogram, WSS- and
silhouette curve), results of the reduction analysis,
heat map of the distances between the observations and SOM nodes as
well as the self-organizing map training process. It is also possible to plot
the first two variables of the clustering data frame, an option which is
attractive, if the clustering of reduction analysis was performed.
}
\references{
Wehrens R, Kruisselbrink J. Flexible self-organizing maps in kohonen 3.0.
J Stat Softw (2018) 87:1–18. doi:10.18637/jss.v087.i07

Kassambara A, Mundt F. factoextra: Extract and Visualize the Results
of Multivariate Data Analyses. (2020) Available
at: https://cran.r-project.org/web/packages/factoextra/index.html

Galili T. dendextend: an R package for visualizing, adjusting and
comparing trees of hierarchical clustering.
Bioinformatics (2015) 31:3718–20. doi:10.1093/bioinformatics/btv428

McInnes L, Healy J, Melville J. UMAP: Uniform Manifold Approximation
and Projection for Dimension Reduction. (2018) Available
at: https://arxiv.org/abs/1802.03426v3

Belyadi H, Haghighat A, Nguyen H, Guerin A-J. IOP Conference Series:
Earth and Environmental Science Determination of Optimal Epsilon (Eps)
Value on DBSCAN Algorithm to Clustering Data on Peatland Hotspots in
Sumatra Related content EPS conference comes to London-EPS rewards
quasiparticle research-EP. IOP Conf Ser Earth Environ Sci (2016) 31:
doi:10.1088/1755-1315/31/1/012012

Rousseeuw PJ. Silhouettes: A graphical aid to the interpretation and
validation of cluster analysis. J Comput Appl Math (1987) 20:53–65.
doi:10.1016/0377-0427(87)90125-7
}
