% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_oop.R
\name{plot.importance}
\alias{plot.importance}
\title{Plot feature importance as a scatter or bar plot.}
\usage{
\method{plot}{importance}(
  x,
  type = c("scatter", "bar"),
  fill_color = "cornsilk3",
  label = TRUE,
  txt_size = 2.75,
  signif_digits = 2,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  cust_theme = ggplot2::theme_classic()
)
}
\arguments{
\item{x}{an importance object.}

\item{type}{type of the plot: scatter or bar. Defaults to scatter.}

\item{fill_color}{fill color for the bars or points.}

\item{label}{logical, should the points be labeled with the importance
stat value?}

\item{txt_size}{label text size.}

\item{signif_digits}{significant digits for rounding of the statistic value.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{cust_theme}{a ggplot theme.}
}
\value{
a ggplot bar or scatter plot.
}
\description{
Generates a bar plot with the importance statistic for
the clustering variables. The importance statistic is the difference in the
clustering variance fraction between the original clustering structure and
the clustering objects with the given variable reshuffled randomly.
}
