% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_analysis_oop.R
\name{plot.red_analysis}
\alias{plot.red_analysis}
\title{Plot features of a red_analysis object.}
\usage{
\method{plot}{red_analysis}(
  x,
  type = c("component_tbl", "scores", "loadings", "scree"),
  label_points = TRUE,
  cust_theme = ggplot2::theme_classic(),
  segment_color = "steelblue",
  ...
)
}
\arguments{
\item{x}{a \code{red_analysis} object, created with \code{\link{reduce_data}}.}

\item{type}{plot type:
'component_tbl' or 'score' present the scores for particular observations in
a scatter plot.
'loadings' plot the variable PCA loadings as a scatter plot.
'scree' plots the percentage of component's variances as a line plot.}

\item{label_points}{logical, should the variable names be displayed in the
plot? Valid only for the PCA loadings plot.}

\item{cust_theme}{a ggplot plot theme.}

\item{segment_color}{color of the lines presented in the PCA loading plot.}

\item{...}{extra arguments passed to \code{\link{plot_point}}.}
}
\value{
a ggplot object.
}
\description{
Plots the component table, loadings table - in both cases the
first two components/dimensions in form of scatter plots -  or generates
a scree plot of the variance percentages associated with
the components/dimensions.
}
\details{
The loadings table plot is available only for the PCA \code{red_analysis}
objects.
}
