% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_graphs.R
\name{plot_clust_hm}
\alias{plot_clust_hm}
\title{Plot levels of clustering features in a heat map.}
\usage{
plot_clust_hm(
  x_object,
  y_object = NULL,
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = "Sample",
  fill_lab = "Feature level",
  cust_theme = ggplot2::theme_classic(),
  discrete_fill = FALSE
)
}
\arguments{
\item{x_object}{a clust_analysis or combi_analysis object, specifies
clustering of the observations.}

\item{y_object}{a clust_analysis or combi_analysis object, specifies
clustering of the features. Optional.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{x_lab}{x axis title.}

\item{fill_lab}{fill scale title.}

\item{cust_theme}{a ggplot theme.}

\item{discrete_fill}{logical, force a discrete fill scale?}
}
\description{
Generates a heat map of the clustering features, cluster
assignment is indicated by the plot facetting.
}
