% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_knn_distance}
\alias{plot_knn_distance}
\title{Plot the mean distance to k-nearest neighbors.}
\usage{
plot_knn_distance(
  diss_obj,
  k,
  eps = NULL,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  cust_theme = ggplot2::theme_classic()
)
}
\arguments{
\item{diss_obj}{a dissimilarity object (e.g. 'dist' class).}

\item{k}{the k number of the nearest neighbors.}

\item{eps}{the distance to be presented in the plot as a horizontal dashed
line. If NULL, the line is hidden.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{cust_theme}{custom plot theme, a ggplot2 theme object.}
}
\value{
A ggplot object.
}
\description{
Plots the sorted (ascending) distances to k-nearest neighbors
(kNN) for each observation in the provided dissimilarity object.
}
\details{
Internally, the mean kNN distances are calculated with the
\code{\link[dbscan]{kNNdist}} function.
}
\references{
Hahsler M, Piekenbrock M, Doran D. Dbscan: Fast density-based clustering
with R. J Stat Softw (2019) 91:1–30. doi:10.18637/jss.v091.i01

Belyadi H, Haghighat A, Nguyen H, Guerin A-J. IOP Conference Series:
Earth and Environmental Science Determination of Optimal Epsilon (Eps)
Value on DBSCAN Algorithm to Clustering Data on Peatland Hotspots in
Sumatra Related content EPS conference comes to London-EPS rewards
quasiparticle research-EP. IOP Conf Ser Earth Environ Sci (2016) 31:
doi:10.1088/1755-1315/31/1/012012
}
