% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_som}
\alias{plot_som}
\title{Plot diagnostic plots for the self-organizing map.}
\usage{
plot_som(kohonen_object)
}
\arguments{
\item{kohonen_object}{a 'kohonen' class object.
See: \code{\link[kohonen]{som}} for details.}
}
\value{
a list of non-editable ggplot objects.
}
\description{
Generates a set of diagnostic plots for the 'kohonen' class
object as specified by \code{\link[kohonen]{plot.kohonen}}.
}
\details{
The plots are wrapped as a list of (non-editable) ggplot objects.
}
