% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi_supervised.R
\name{predict.clust_analysis}
\alias{predict.clust_analysis}
\alias{predict.combi_analysis}
\title{Semi-supervised clustering.}
\usage{
\method{predict}{clust_analysis}(object, newdata = NULL, type = c("class", "propagation"), ...)

\method{predict}{combi_analysis}(object, newdata = NULL, type = c("class", "propagation"), ...)
}
\arguments{
\item{object}{an object.}

\item{newdata}{a numeric data frame, matrix or a red_analysis object.
If NULL (default), the bare cluster assignment table is returned.}

\item{type}{type of the projection: simple observation matching
('class', default) or kNN label propagation ('propagation').}

\item{...}{extra arguments passed to \code{\link{propagate}}.}
}
\value{
a \code{\link{clust_analysis}} object.
}
\description{
Projects the cluster assignment onto new data using simple
observation matching or a k-nearest neighbor (kNN) label propagation
algorithm.
}
\details{
For the implementation details, see: \code{\link{propagate}}.
The default distance metric is extracted from the `clust_analysis` object.
For `combi_analysis` objects, the default distance metric is the distance
between observations (not nodes!).
The cluster projection is done on the top level, i.e. takes into account the
final assignment of the observations to the clusters and ignoring
the SOM nodes.
}
\references{
Leng M, Wang J, Cheng J, Zhou H, Chen X. Adaptive
semi-supervised clustering algorithm with label propagation.
J Softw Eng (2014) 8:14–22. doi:10.3923/jse.2014.14.22
}
