% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis_oop.R
\name{print.clust_analysis}
\alias{print.clust_analysis}
\title{Printing of a clust_analysis object.}
\usage{
\method{print}{clust_analysis}(x, ...)
}
\arguments{
\item{x}{a clust_analysis object.}

\item{...}{extra arguments, currently none.}
}
\description{
Prints a clust_analysis object.
}
