% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{red_analysis}
\alias{red_analysis}
\title{Construct a red_analysis class object.}
\usage{
red_analysis(x)
}
\arguments{
\item{x}{a named list, see details.}
}
\description{
Constructs a red_analysis class object given a list storing
results of a dimensionality reduction analysis.
}
\details{
A named list with the following elements ir required:
red_obj (analysis output),
red_fun (name of the reduction function),
component_tbl (component values for the observations),
loadings (variable loadings, relevant only for PCA),
data (a quosure calling the original data set).
}
