% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{som_cluster}
\alias{som_cluster}
\title{Self-organizing maps.}
\usage{
som_cluster(
  data,
  distance_method = "euclidean",
  xdim = 5,
  ydim = 4,
  topo = c("hexagonal", "rectangular"),
  neighbourhood.fct = c("gaussian", "bubble"),
  toroidal = FALSE,
  seed = 1234,
  ...
)
}
\arguments{
\item{data}{a numeric data frame or matrix or a red_analysis object.}

\item{distance_method}{name of the distance metric, see:
\code{\link{get_kernel_info}}.}

\item{xdim}{x dimension of the SOM grid,
see: \code{\link[kohonen]{somgrid}} for details.}

\item{ydim}{y dimension of the SOM grid,
#' see: \code{\link[kohonen]{somgrid}} for details.}

\item{topo}{SOM grid topology, see: \code{\link[kohonen]{somgrid}}
for details. 'hexagonal' for default.}

\item{neighbourhood.fct}{neighborhood function, 'gaussian' for default.}

\item{toroidal}{logical, should toroidal grid be used?}

\item{seed}{initial setting of the random number generator.}

\item{...}{extra arguments passed to \code{\link[kohonen]{som}}.}
}
\value{
an object of the class \code{\link{clust_analysis}}.
}
\description{
Performs self-organizing map (SOM) clustering of a numeric data
frame, matrix or the results of a reduction analysis.
}
\details{
Technically, a wrapper around \code{\link[kohonen]{som}}. If a
red_analysis object is provided as the data argument, the observation
component/score table is subjected to clustering. Note, in order to make use
of the full set of distance measures, the package 'somKernels' need to be
installed and loaded.
}
\references{
Kohonen T. Self-Organizing Maps. Berlin, Heidelberg:
Springer Berlin Heidelberg (1995). doi:10.1007/978-3-642-97610-0

Wehrens R, Kruisselbrink J. Flexible self-organizing maps in kohonen
3.0. J Stat Softw (2018) 87:1–18. doi:10.18637/jss.v087.i07
}
