% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{var}
\alias{var}
\alias{var.default}
\title{Object's variance}
\usage{
var(x, ...)

\method{var}{default}(x, ...)
}
\arguments{
\item{x}{an object. For the default method a numeric vector, matrix
or a data frame.}

\item{...}{extra arguments passed to methods, e.g. \code{\link[stats]{var}}.}
}
\description{
Computes variance statistic specific for the given object.
}
\details{
The default `var()` method is a wrapper around
\code{\link[stats]{var}}.
}
