% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_analysis_oop.R
\name{var.red_analysis}
\alias{var.red_analysis}
\alias{summary.red_analysis}
\title{Variance and summary for a red_analysis object.}
\usage{
\method{var}{red_analysis}(x, ...)

\method{summary}{red_analysis}(object, ...)
}
\arguments{
\item{x}{a \code{red_analysis} object.}

\item{...}{extra arguments, currently none.}

\item{object}{a \code{red_analysis} object.}
}
\value{
\code{var()} returns a data frame with components'
variances, \code{summary()} returns a set of summary statistic specific for
the wrapped dimensionality reduction function.
}
\description{
Variance associated with the
components and statistic summary for \code{red_analysis} class objects.
}
